Validate the local library

Checking if installed packages have correct versions taking into account all DESCRIPTION files requirements. Moreover identifying which packages are newest releases. Optionally we could add life duration and CRAN check page status for each package.

lib_validate(
  lib.loc = NULL,
  fields = c("Depends", "Imports", "LinkingTo"),
  lifeduration = FALSE,
  checkred = character(0),
  repos = "https://cran.rstudio.com/",
  cores = max(parallel::detectCores() - 1, 1),
  parallel = "auto"
)

Arguments

lib.loc
character. Default: NULL

fields
character vector with possible values c("Depends", "Imports", "LinkingTo", "Suggests"). Default: c("Depends", "Imports", "LinkingTo")

lifeduration
logical if to add life duration column, might take some time. Default: FALSE

checkred
character if to add R CRAN check page status, any of c("ERROR", "FAIL", "WARN", "NOTE"). Default character(0)

repos
character the base URL of the repositories to use. Default https://cran.rstudio.com/.

cores
numeric number of cores to use when lifeduration or checkred is activated. Default: max(parallel::detectCores() - 1, 1)

parallel
character type of parallel calculations when lifeduration or checkred is activated, c("auto", "none", "mclapply", "parLapply"). Default: "auto"

Value

data.frame with 5/6/7 columns.

Package

character package names.

Version.expected.min

character expected by DESCRIPTION files minimal version. "" means not specified so the newest version.

Version.have

character installed package version.

version_status

numeric -1/0/1 which comes from utils::compareVersion function. 0 means that we have the same version as required by DESCRIPTION files. -1 means we have too low version installed, this is an error. 1 means we have higher version.

newest

logical if the installed version is the newest one.

checkred

(Optional) logical if the newest package contains any errors or warnings on CRAN check page.

life_duration

(Optional) integer number of days package was released.

Note

Version.expected.min column not count packages which are not a dependency for any package, so could not be find in DESCRIPTION files. When turn on the lifeduration and/or checkred options, calculations might be time consuming. Please as a courtesy to the R CRAN, don't overload their server by constantly using this function with lifeduration or checkred turned on. Results are NOT cached with memoise package, when parallel::mclapply function is used. For lifeduration and checkred options there is used the parallel::mclapply function or parallel::parLapply, cores on default is set to all cores minus one. cores argument has lower priority than not NULL getOption("mc.cores")or getOption("cl.cores"). Remember that parallel::mclapply under Windows works like the regular lapply function, changing cores will be neutral for Windows machines.

Examples

lib_validate()
#>         Package Version.expected.min Version.have version_status newest
#> 1    KernSmooth                 <NA>      2.23-20              1   TRUE
#> 2          MASS                            7.3-54              1   TRUE
#> 3        Matrix                             1.3-3              1  FALSE
#> 4            R6                2.2.0        2.5.0              1   TRUE
#> 5          Rcpp               0.11.0        1.0.6              1  FALSE
#> 6       askpass                               1.1              1   TRUE
#> 7     base64enc                             0.1-3              1   TRUE
#> 8          boot                 <NA>       1.3-28              1   TRUE
#> 9          brew                             1.0-6              1   TRUE
#> 10         brio                             1.1.2              1   TRUE
#> 11        bslib                           0.2.5.1              1   TRUE
#> 12       cachem                             1.0.5              1   TRUE
#> 13        callr                3.6.0        3.7.0              1   TRUE
#> 14        class                 <NA>       7.3-19              1   TRUE
#> 15          cli                2.4.0        3.0.1              1   TRUE
#> 16        clipr                0.3.0        0.7.1              1   TRUE
#> 17      cluster                 <NA>        2.1.2              1   TRUE
#> 18    codetools                 <NA>       0.2-18              1   TRUE
#> 19   commonmark                               1.7              1   TRUE
#> 20        cpp11                0.2.1        0.3.1              1   TRUE
#> 21       crayon                1.3.4        1.4.1              1   TRUE
#> 22  credentials                1.2.1        1.3.0              1  FALSE
#> 23         curl                3.0.0        4.3.2              1   TRUE
#> 24         desc                1.3.0        1.3.0              0   TRUE
#> 25     devtools                 <NA>        2.4.2              1   TRUE
#> 26      diffobj                             0.3.4              1   TRUE
#> 27       digest                            0.6.27              1   TRUE
#> 28      downlit                        0.2.1.9000              1  FALSE
#> 29     ellipsis                0.3.2        0.3.2              0   TRUE
#> 30     evaluate                 0.13         0.14              1   TRUE
#> 31        fansi                0.4.0        0.5.0              1   TRUE
#> 32      fastmap                             1.1.0              1   TRUE
#> 33      foreign                 <NA>       0.8-81              1   TRUE
#> 34           fs                1.5.0        1.5.0              0   TRUE
#> 35         gert                1.0.2        1.3.1              1   TRUE
#> 36           gh                1.2.0        1.3.0              1   TRUE
#> 37     gitcreds                             0.1.1              1   TRUE
#> 38         glue                1.3.0        1.4.2              1   TRUE
#> 39        highr                               0.9              1   TRUE
#> 40    htmltools                0.5.1      0.5.1.1              1   TRUE
#> 41         httr                1.4.2        1.4.2              0   TRUE
#> 42          ini                             0.3.1              1   TRUE
#> 43    jquerylib                0.1.3        0.1.4              1   TRUE
#> 44     jsonlite                             1.7.2              1   TRUE
#> 45        knitr                 1.22         1.33              1   TRUE
#> 46      lattice                           0.20-44              1   TRUE
#> 47    lifecycle                1.0.0        1.0.0              0   TRUE
#> 48     magrittr                  1.5        2.0.1              1   TRUE
#> 49     markdown                               1.1              1   TRUE
#> 50      memoise                2.0.0        2.0.0              0   TRUE
#> 51         mgcv                 <NA>       1.8-35              1  FALSE
#> 52         mime                  0.3         0.11              1   TRUE
#> 53         nlme               3.1-64      3.1-152              1   TRUE
#> 54         nnet                 <NA>       7.3-16              1   TRUE
#> 55      openssl                1.4.1        1.4.4              1   TRUE
#> 56      packrat                 <NA>        0.6.0              1   TRUE
#> 57         pacs                 <NA>       0.2.10              1  FALSE
#> 58       pillar                1.6.0        1.6.2              1   TRUE
#> 59     pkgbuild                1.2.0        1.2.0              0   TRUE
#> 60    pkgconfig                             2.0.3              1   TRUE
#> 61      pkgdown                 <NA>   1.6.1.9001              1  FALSE
#> 62      pkgload                1.2.1        1.2.1              0   TRUE
#> 63       praise                             1.0.0              1   TRUE
#> 64  prettyunits                             1.1.1              1   TRUE
#> 65     processx                3.5.0        3.5.2              1   TRUE
#> 66           ps                1.3.4        1.6.0              1   TRUE
#> 67        purrr                0.3.3        0.3.4              1   TRUE
#> 68         ragg                             1.1.3              1   TRUE
#> 69     rappdirs                             0.3.3              1   TRUE
#> 70    rcmdcheck                1.3.3        1.3.3              0   TRUE
#> 71     rematch2                             2.1.2              1   TRUE
#> 72      remotes                2.3.0        2.4.0              1   TRUE
#> 73        rlang               0.4.10       0.4.11              1   TRUE
#> 74    rmarkdown             1.1.9007          2.9              1  FALSE
#> 75     roxygen2                7.1.1        7.1.1              0   TRUE
#> 76        rpart                 <NA>       4.1-15              1   TRUE
#> 77    rprojroot                  1.2        2.0.2              1   TRUE
#> 78   rstudioapi                 0.13         0.13              0   TRUE
#> 79    rversions                2.0.2        2.1.1              1   TRUE
#> 80         sass                0.4.0        0.4.0              0   TRUE
#> 81  sessioninfo                1.1.1        1.1.1              0   TRUE
#> 82      spatial                 <NA>       7.3-14              1   TRUE
#> 83      stringi                1.1.7        1.7.3              1   TRUE
#> 84      stringr                1.2.0        1.4.0              1   TRUE
#> 85     survival                 <NA>       3.2-11              1   TRUE
#> 86          sys                  2.1          3.4              1   TRUE
#> 87  systemfonts                1.0.0        1.0.2              1   TRUE
#> 88     testthat                3.0.2        3.0.4              1   TRUE
#> 89  textshaping                0.3.0        0.3.5              1   TRUE
#> 90       tibble                             3.1.3              1   TRUE
#> 91      tinytex                 0.31         0.32              1  FALSE
#> 92      usethis                2.0.1        2.0.1              0   TRUE
#> 93         utf8                1.1.0        1.2.2              1   TRUE
#> 94        vctrs                0.3.8        0.3.8              0   TRUE
#> 95        waldo                0.2.4        0.2.5              1   TRUE
#> 96      whisker                               0.4              1   TRUE
#> 97        withr                2.4.1        2.4.2              1   TRUE
#> 98         xfun                 0.23         0.24              1  FALSE
#> 99         xml2                1.3.1        1.3.2              1   TRUE
#> 100       xopen                             1.0.0              1   TRUE
#> 101        yaml               2.1.19        2.2.1              1   TRUE
#> 102         zip                2.1.0        2.2.0              1   TRUE
#> 103           R                4.0.0        4.1.0              1     NA