Checking the R CRAN package check page status

using package R CRAN check page to validate if there are ANY errors and/or fails and/or warnings and/or notes.

pac_checkred(pac, scope = c("ERROR", "FAIL"), flavors = NULL)

Arguments

pac

character a package name.

scope

character vector scope of the check, accepted values c("ERROR", "FAIL", "WARN", "NOTE"). Default c("ERROR", "FAIL")

flavors

character vector of CRAN machines to consider, which might be retrieved with pacs::cran_flavors()$Flavor. By default all CRAN machines are considered, NULL value. Default NULL

Value

logical if the package fail under specified criteria.

Note

Results are cached for 1 hour with memoise package. If you need to check many packages at once then is recommended usage of pacs::checked_packages. The used repository https://cran.rstudio.com/. Please as a courtesy to the R CRAN, don't overload their server by constantly using this function.

Examples

if (FALSE) {
pac_checkred("dplyr")
pac_checkred("dplyr", scope = c("ERROR"))
pac_checkred("dplyr",
  scope = c("ERROR", "FAIL", "WARN"),
  flavors = c(
    "r-devel-linux-x86_64-debian-clang",
    "r-devel-linux-x86_64-debian-gcc"
  )
)
}