Compare DESCRIPTION files dependencies between specific CRAN packages versions

using the remote github CRAN mirror to compare DESCRIPTION files dependencies between specific packages versions.

pac_compare_versions(
  pac,
  old = NULL,
  new = NULL,
  fields = c("Imports", "Depends", "LinkingTo"),
  lib.loc = NULL,
  repos = "https://cran.rstudio.com/"
)

Arguments

pac
character a package name.

old
character an old version of package, default local version. Default: NULL

new
character a new version of package, default newest version. Default: NULL

fields
character a vector with possible values c("Depends", "Imports", "LinkingTo", "Suggests"). Default: c("Depends", "Imports", "LinkingTo")

lib.loc
character. Default: NULL

repos
character the base URL of the CRAN repository to use. Default "https://cran.rstudio.org"

Value

data.frame with 4 columns.

Package

character package names.

Version.OLD

character versions of dependencies required by an old package version.

Version.NEW

character versions of dependencies required by a new package version.

version_status

numeric -1/0/1 which comes from utils::compareVersion function. 0 means that we have the same version as required by DESCRIPTION files. -1 means we have too low version installed, this is an error. 1 means we have higher version.

Examples

if (FALSE) {
pac_compare_versions("memoise", "0.2.1", "2.0.0")
pac_compare_versions("memoise", "0.2.1")
}