package DESCRIPTION file

package DESCRIPTION file taken locally or if remotely then from GITHUB CRAN mirror or CRAN website.

pac_description(pac, version = NULL, at = NULL, local = FALSE, lib.loc = NULL)

Arguments

pac
character a package name.

version
character package version. Default: NULL

at
Date. Default: NULL

local
logical if to use local library. Default: FALSE

lib.loc
character used optionally when local is equal TRUE. Default: NULL

Value

list with names proper for DESCRIPTION file fields.

Examples

pac_description("dplyr", version = "0.8.0")
#> $Type
#> [1] "Package"
#> 
#> $Package
#> [1] "dplyr"
#> 
#> $Title
#> [1] "A Grammar of Data Manipulation"
#> 
#> $Version
#> [1] "0.8.0"
#> 
#> $`Authors@R`
#> [1] "c(\nperson(\"Hadley\", \"Wickham\", , \"hadley@rstudio.com\", c(\"aut\", \"cre\"), comment = c(ORCID = \"0000-0003-4757-117X\")),\nperson(\"Romain\", \"Fran\\u00e7ois\", role = \"aut\", comment = c(ORCID = \"0000-0002-2444-4226\")),\nperson(\"Lionel\", \"Henry\", role = \"aut\"),\nperson(\"Kirill\", \"M\\u00fcller\", role = \"aut\", comment = c(ORCID = \"0000-0002-1416-3412\")),\nperson(\"RStudio\", role = c(\"cph\", \"fnd\"))\n)"
#> 
#> $Description
#> [1] "A fast, consistent tool for working with data frame like objects,\nboth in memory and out of memory."
#> 
#> $License
#> [1] "MIT + file LICENSE"
#> 
#> $URL
#> [1] "http://dplyr.tidyverse.org, https://github.com/tidyverse/dplyr"
#> 
#> $BugReports
#> [1] "https://github.com/tidyverse/dplyr/issues"
#> 
#> $Depends
#> [1] "R (>= 3.1.2)"
#> 
#> $Imports
#> [1] "assertthat (>= 0.2.0), glue (>= 1.1.1), magrittr (>= 1.5),\nmethods, pkgconfig (>= 2.0.1), R6 (>= 2.2.2), Rcpp (>= 1.0.0),\nrlang (>= 0.3.0), tibble (>= 2.0.0), tidyselect (>= 0.2.5),\nutils"
#> 
#> $Suggests
#> [1] "bit64 (>= 0.9.7), callr (>= 3.1.1), covr (>= 3.0.1), DBI (>=\n0.7.14), dbplyr (>= 1.2.0), dtplyr (>= 0.0.2), ggplot2 (>=\n2.2.1), hms (>= 0.4.1), knitr (>= 1.19), Lahman (>= 3.0-1),\nlubridate (>= 1.7.4), MASS, mgcv (>= 1.8.23), microbenchmark\n(>= 1.4.4), nycflights13 (>= 0.2.2), rmarkdown (>= 1.8), RMySQL\n(>= 0.10.13), RPostgreSQL (>= 0.6.2), RSQLite (>= 2.0),\ntestthat (>= 2.0.0), withr (>= 2.1.1), broom (>= 0.5.1), purrr\n(>= 0.3.0), readr (>= 1.3.1), crayon (>= 1.3.4)"
#> 
#> $LinkingTo
#> [1] "BH (>= 1.58.0-1), plogr (>= 0.1.10), Rcpp (>= 1.0.0)"
#> 
#> $VignetteBuilder
#> [1] "knitr"
#> 
#> $Encoding
#> [1] "UTF-8"
#> 
#> $LazyData
#> [1] "yes"
#> 
#> $RoxygenNote
#> [1] "6.1.1"
#> 
#> $NeedsCompilation
#> [1] "yes"
#> 
#> $Packaged
#> [1] "2019-02-12 16:18:43 UTC; romain"
#> 
#> $Author
#> [1] "Hadley Wickham [aut, cre] (<https://orcid.org/0000-0003-4757-117X>),\nRomain François [aut] (<https://orcid.org/0000-0002-2444-4226>),\nLionel Henry [aut],\nKirill Müller [aut] (<https://orcid.org/0000-0002-1416-3412>),\nRStudio [cph, fnd]"
#> 
#> $Maintainer
#> [1] "Hadley Wickham <hadley@rstudio.com>"
#> 
#> $Repository
#> [1] "CRAN"
#> 
#> $`Date/Publication`
#> [1] "2019-02-14 22:22:14 UTC"
#> 
pac_description("dplyr", at = as.Date("2019-02-01"))
#> $Type
#> [1] "Package"
#> 
#> $Package
#> [1] "dplyr"
#> 
#> $Title
#> [1] "A Grammar of Data Manipulation"
#> 
#> $Version
#> [1] "0.7.8"
#> 
#> $`Authors@R`
#> [1] "c(\nperson(\"Hadley\", \"Wickham\", , \"hadley@rstudio.com\", c(\"aut\", \"cre\"), comment = c(ORCID = \"0000-0003-4757-117X\")),\nperson(\"Romain\", \"Fran\\u00e7ois\", role = \"aut\", comment = c(ORCID = \"0000-0002-2444-4226\")),\nperson(\"Lionel\", \"Henry\", role = \"aut\"),\nperson(\"Kirill\", \"M\\u00fcller\", role = \"aut\", comment = c(ORCID = \"0000-0002-1416-3412\")),\nperson(\"RStudio\", role = c(\"cph\", \"fnd\"))\n)"
#> 
#> $Description
#> [1] "A fast, consistent tool for working with data frame like objects,\nboth in memory and out of memory."
#> 
#> $License
#> [1] "MIT + file LICENSE"
#> 
#> $URL
#> [1] "http://dplyr.tidyverse.org, https://github.com/tidyverse/dplyr"
#> 
#> $BugReports
#> [1] "https://github.com/tidyverse/dplyr/issues"
#> 
#> $Depends
#> [1] "R (>= 3.1.2)"
#> 
#> $Imports
#> [1] "assertthat (>= 0.2.0), bindrcpp (>= 0.2.0.9000), glue (>=\n1.1.1), magrittr (>= 1.5), methods, pkgconfig (>= 2.0.1), R6\n(>= 2.2.2), Rcpp (>= 0.12.19), rlang (>= 0.3.0), tibble (>=\n1.4.2), tidyselect (>= 0.2.3), utils"
#> 
#> $Suggests
#> [1] "bit64 (>= 0.9.7), callr, covr (>= 3.0.1), DBI (>= 0.7.14),\ndbplyr (>= 1.2.0), dtplyr (>= 0.0.2), ggplot2 (>= 2.2.1), hms\n(>= 0.4.1), knitr (>= 1.19), Lahman (>= 3.0-1), lubridate,\nMASS, mgcv (>= 1.8.23), microbenchmark (>= 1.4.4), nycflights13\n(>= 0.2.2), rmarkdown (>= 1.8), RMySQL (>= 0.10.13),\nRPostgreSQL (>= 0.6.2), RSQLite (>= 2.0), testthat (>= 2.0.0),\nwithr (>= 2.1.1)"
#> 
#> $LinkingTo
#> [1] "BH (>= 1.58.0-1), bindrcpp (>= 0.2.0.9000), plogr (>=\n0.1.10), Rcpp (>= 0.12.15)"
#> 
#> $VignetteBuilder
#> [1] "knitr"
#> 
#> $Encoding
#> [1] "UTF-8"
#> 
#> $LazyData
#> [1] "yes"
#> 
#> $RoxygenNote
#> [1] "6.1.0"
#> 
#> $NeedsCompilation
#> [1] "yes"
#> 
#> $Packaged
#> [1] "2018-11-09 21:23:40 UTC; kirill"
#> 
#> $Author
#> [1] "Hadley Wickham [aut, cre] (<https://orcid.org/0000-0003-4757-117X>),\nRomain François [aut] (<https://orcid.org/0000-0002-2444-4226>),\nLionel Henry [aut],\nKirill Müller [aut] (<https://orcid.org/0000-0002-1416-3412>),\nRStudio [cph, fnd]"
#> 
#> $Maintainer
#> [1] "Hadley Wickham <hadley@rstudio.com>"
#> 
#> $Repository
#> [1] "CRAN"
#> 
#> $`Date/Publication`
#> [1] "2018-11-10 07:30:02 UTC"
#>