CRAN package health state at a specific Date or for a specific version

using CRAN website to get a package version/versions used at a specific Date interval. A healthy package was published for more than x days, where default is 14 days. CRAN team gives around one/two week to resolved a package which gave errors under the check page.

pac_health(
  pac,
  version = NULL,
  at = NULL,
  limit = 14,
  scope = c("ERROR", "FAIL"),
  flavors = NULL,
  lib.loc = .libPaths(),
  repos = "https://cran.rstudio.com/",
  source = c("crandb", "cran")
)

Arguments

pac

character a package name.

version

character package version, By default the newest version is taken. Default: NULL

at

Date old version of package. Default: NULL

limit

numeric at least days to treat as healthy, ">=limit". Default: 14

scope

character vector scope of R CRAN check pages statuses to consider, any of c("ERROR", "FAIL", "WARN", "NOTE"). Default c("ERROR", "FAIL")

flavors

character vector of CRAN machines to consider, which might be retrieved with pacs::cran_flavors()$Flavor. By default all CRAN machines are considered, NULL value. Default NULL

lib.loc

character vector. Is omitted for non NULL version. Default: .libPaths()

repos

character the base CRAN URL of the repository to use. Default "https://cran.rstudio.org"

source

character one of c("crandb", "cran"). Using the MEATCRAN DB or the direct web page download from CRAN. Default: "crandb"

Value

logical if a package is healthy.

Note

Function will scrap two/tree CRAN URLS. Works only with CRAN packages. The newest release are checked for warnings/errors on R CRAN check page. Please as a courtesy to the R CRAN, don't overload their server by constantly using this function. Results are cached for 30 minutes with memoise package, memory cache. The crandb R packages database is a part of METACRAN project, source: Csárdi G, Salmon M (2022). pkgsearch: Search and Query CRAN R Packages. https://github.com/r-hub/pkgsearch, https://r-hub.github.io/pkgsearch/.

Examples

if (FALSE) {
pac_health("memoise")
pac_health("dplyr", version = "0.8.0", limit = 14)
pac_health("dplyr", at = as.Date("2019-02-14"))
pac_health("dplyr", limit = 14, scope = c("ERROR", "FAIL"))
}