Checking if a packge version is the most recent one

checking if a package version is the most recent one, by default the installed version is compared.

pac_islast(
  pac,
  version = NULL,
  lib.loc = NULL,
  repos = c("https://bioconductor.org/packages/3.13/bioc",
    "https://bioconductor.org/packages/3.13/data/annotation",
    "https://bioconductor.org/packages/3.13/data/experiment",
    "https://bioconductor.org/packages/3.13/workflows",
    "https://bioconductor.org/packages/3.13/books", "https://cloud.r-project.org")
)

Arguments

pac
character a package name.

version
character package version, by default the installed version is taken. Default: NULL

lib.loc
character vector. Is omitted for non NULL version. Default: NULL

repos
character vector base URLs of the repositories to use. By default checking CRAN and Bioconductor. Default pacs::biocran_repos()

Value

logical if a package is inside ropositories.

Note

Results are cached for 1 hour with memoise package.

Examples

if (FALSE) {
pac_islast("memoise")
pac_islast("dplyr", version = "1.0.0")
pac_lifeduration("S4Vectors")
pac_lifeduration("S4Vectors", version = pac_last("S4Vectors"))
}