Package version life duration at specific Date or for a specific version

using CRAN website to get a package life duration for certain version or at a specific Date.

pac_lifeduration(
  pac,
  version = NULL,
  at = NULL,
  lib.loc = .libPaths(),
  repos = biocran_repos()
)

Arguments

pac

character a package name.

version

character package version, By default the newest version is taken. Default: NULL

at

Date old version of package. Default: NULL

lib.loc

character vector. Is omitted for non NULL version. Default: .libPaths()

repos

character vector base URLs of the repositories to use. By default checking CRAN and newest Bioconductor. Default pacs::biocran_repos()

Value

difftime, number of days package version was the newest one.

Note

Function will scrap two github CRAN mirror and CRAN URL. Works mainly with CRAN packages. Please as a courtesy to the R CRAN, don't overload their server by constantly using this function. Results are cached for 30 minutes with memoise package, memory cache.

Examples

if (FALSE) {
pac_lifeduration("memoise")
pac_lifeduration("dplyr", version = "0.8.0")
pac_lifeduration("dplyr", at = as.Date("2019-02-14"))
# For Bioconductor packages will work only for the newest per R version and installed packages.
pac_lifeduration("S4Vectors")
}