Package NAMESPACE file

package NAMESPACE file taken locally or remotely from GITHUB CRAN mirror or CRAN website.

pac_namespace(
  pac,
  version = NULL,
  at = NULL,
  local = FALSE,
  lib.loc = NULL,
  repos = "https://cran.rstudio.com/"
)

Arguments

pac
character a package name.

version
character package version. Default: NULL

at
Date. Default: NULL

local
logical if to use local library. Default: FALSE

lib.loc
character used optionally when local is equal TRUE. Default: NULL

repos
character the base URL of the repository to use. Used only for the validation. Default https://cran.rstudio.com/

Value

list with names proper for NAMESPACE file, the same as format as returned by base::parseNamespaceFile.

Note

Results are cached for 1 hour with memoise package. This function is mainly built under source code from base::parseNamespaceFile.

Examples

pac_namespace("dplyr", version = "0.8.0")
#> $imports
#> $imports[[1]]
#> [1] "rlang"
#> 
#> $imports[[2]]
#> $imports[[2]][[1]]
#> [1] "R6"
#> 
#> $imports[[2]][[2]]
#> [1] "R6Class"
#> 
#> 
#> $imports[[3]]
#> $imports[[3]][[1]]
#> [1] "Rcpp"
#> 
#> $imports[[3]][[2]]
#> [1] "Rcpp.plugin.maker"
#> 
#> 
#> $imports[[4]]
#> $imports[[4]][[1]]
#> [1] "Rcpp"
#> 
#> $imports[[4]][[2]]
#> [1] "cppFunction"
#> 
#> 
#> $imports[[5]]
#> $imports[[5]][[1]]
#> [1] "assertthat"
#> 
#> $imports[[5]][[2]]
#> [1] "on_failure<-"
#> 
#> 
#> $imports[[6]]
#> $imports[[6]][[1]]
#> [1] "assertthat"
#> 
#> $imports[[6]][[2]]
#> [1] "assert_that"
#> 
#> 
#> $imports[[7]]
#> $imports[[7]][[1]]
#> [1] "assertthat"
#> 
#> $imports[[7]][[2]]
#> [1] "is.flag"
#> 
#> 
#> $imports[[8]]
#> $imports[[8]][[1]]
#> [1] "glue"
#> 
#> $imports[[8]][[2]]
#> [1] "glue"
#> 
#> 
#> $imports[[9]]
#> $imports[[9]][[1]]
#> [1] "magrittr"
#> 
#> $imports[[9]][[2]]
#> [1] "%>%"
#> 
#> 
#> $imports[[10]]
#> $imports[[10]][[1]]
#> [1] "methods"
#> 
#> $imports[[10]][[2]]
#> [1] "is"
#> 
#> 
#> $imports[[11]]
#> $imports[[11]][[1]]
#> [1] "pkgconfig"
#> 
#> $imports[[11]][[2]]
#> [1] "get_config"
#> 
#> 
#> $imports[[12]]
#> $imports[[12]][[1]]
#> [1] "rlang"
#> 
#> $imports[[12]][[2]]
#> [1] "dots_n"
#> 
#> 
#> $imports[[13]]
#> $imports[[13]][[1]]
#> [1] "stats"
#> 
#> $imports[[13]][[2]]
#> [1] "lag"
#> 
#> 
#> $imports[[14]]
#> $imports[[14]][[1]]
#> [1] "stats"
#> 
#> $imports[[14]][[2]]
#> [1] "setNames"
#> 
#> 
#> $imports[[15]]
#> $imports[[15]][[1]]
#> [1] "stats"
#> 
#> $imports[[15]][[2]]
#> [1] "update"
#> 
#> 
#> $imports[[16]]
#> $imports[[16]][[1]]
#> [1] "tibble"
#> 
#> $imports[[16]][[2]]
#> [1] "add_row"
#> 
#> 
#> $imports[[17]]
#> $imports[[17]][[1]]
#> [1] "tibble"
#> 
#> $imports[[17]][[2]]
#> [1] "as_data_frame"
#> 
#> 
#> $imports[[18]]
#> $imports[[18]][[1]]
#> [1] "tibble"
#> 
#> $imports[[18]][[2]]
#> [1] "as_tibble"
#> 
#> 
#> $imports[[19]]
#> $imports[[19]][[1]]
#> [1] "tibble"
#> 
#> $imports[[19]][[2]]
#> [1] "data_frame"
#> 
#> 
#> $imports[[20]]
#> $imports[[20]][[1]]
#> [1] "tibble"
#> 
#> $imports[[20]][[2]]
#> [1] "data_frame_"
#> 
#> 
#> $imports[[21]]
#> $imports[[21]][[1]]
#> [1] "tibble"
#> 
#> $imports[[21]][[2]]
#> [1] "frame_data"
#> 
#> 
#> $imports[[22]]
#> $imports[[22]][[1]]
#> [1] "tibble"
#> 
#> $imports[[22]][[2]]
#> [1] "glimpse"
#> 
#> 
#> $imports[[23]]
#> $imports[[23]][[1]]
#> [1] "tibble"
#> 
#> $imports[[23]][[2]]
#> [1] "is_tibble"
#> 
#> 
#> $imports[[24]]
#> $imports[[24]][[1]]
#> [1] "tibble"
#> 
#> $imports[[24]][[2]]
#> [1] "lst"
#> 
#> 
#> $imports[[25]]
#> $imports[[25]][[1]]
#> [1] "tibble"
#> 
#> $imports[[25]][[2]]
#> [1] "lst_"
#> 
#> 
#> $imports[[26]]
#> $imports[[26]][[1]]
#> [1] "tibble"
#> 
#> $imports[[26]][[2]]
#> [1] "new_tibble"
#> 
#> 
#> $imports[[27]]
#> $imports[[27]][[1]]
#> [1] "tibble"
#> 
#> $imports[[27]][[2]]
#> [1] "tbl_sum"
#> 
#> 
#> $imports[[28]]
#> $imports[[28]][[1]]
#> [1] "tibble"
#> 
#> $imports[[28]][[2]]
#> [1] "tibble"
#> 
#> 
#> $imports[[29]]
#> $imports[[29]][[1]]
#> [1] "tibble"
#> 
#> $imports[[29]][[2]]
#> [1] "tribble"
#> 
#> 
#> $imports[[30]]
#> $imports[[30]][[1]]
#> [1] "tibble"
#> 
#> $imports[[30]][[2]]
#> [1] "trunc_mat"
#> 
#> 
#> $imports[[31]]
#> $imports[[31]][[1]]
#> [1] "tibble"
#> 
#> $imports[[31]][[2]]
#> [1] "type_sum"
#> 
#> 
#> $imports[[32]]
#> $imports[[32]][[1]]
#> [1] "tidyselect"
#> 
#> $imports[[32]][[2]]
#> [1] "contains"
#> 
#> 
#> $imports[[33]]
#> $imports[[33]][[1]]
#> [1] "tidyselect"
#> 
#> $imports[[33]][[2]]
#> [1] "ends_with"
#> 
#> 
#> $imports[[34]]
#> $imports[[34]][[1]]
#> [1] "tidyselect"
#> 
#> $imports[[34]][[2]]
#> [1] "everything"
#> 
#> 
#> $imports[[35]]
#> $imports[[35]][[1]]
#> [1] "tidyselect"
#> 
#> $imports[[35]][[2]]
#> [1] "last_col"
#> 
#> 
#> $imports[[36]]
#> $imports[[36]][[1]]
#> [1] "tidyselect"
#> 
#> $imports[[36]][[2]]
#> [1] "matches"
#> 
#> 
#> $imports[[37]]
#> $imports[[37]][[1]]
#> [1] "tidyselect"
#> 
#> $imports[[37]][[2]]
#> [1] "num_range"
#> 
#> 
#> $imports[[38]]
#> $imports[[38]][[1]]
#> [1] "tidyselect"
#> 
#> $imports[[38]][[2]]
#> [1] "one_of"
#> 
#> 
#> $imports[[39]]
#> $imports[[39]][[1]]
#> [1] "tidyselect"
#> 
#> $imports[[39]][[2]]
#> [1] "starts_with"
#> 
#> 
#> $imports[[40]]
#> $imports[[40]][[1]]
#> [1] "utils"
#> 
#> $imports[[40]][[2]]
#> [1] "head"
#> 
#> 
#> $imports[[41]]
#> $imports[[41]][[1]]
#> [1] "utils"
#> 
#> $imports[[41]][[2]]
#> [1] "tail"
#> 
#> 
#> 
#> $exports
#>   [1] "%>%"                 ".data"               "add_count"          
#>   [4] "add_count_"          "add_row"             "add_rownames"       
#>   [7] "add_tally"           "add_tally_"          "all_equal"          
#>  [10] "all_vars"            "anti_join"           "any_vars"           
#>  [13] "arrange"             "arrange_"            "arrange_all"        
#>  [16] "arrange_at"          "arrange_if"          "as.tbl"             
#>  [19] "as.tbl_cube"         "as_data_frame"       "as_tibble"          
#>  [22] "auto_copy"           "bench_tbls"          "between"            
#>  [25] "bind_cols"           "bind_rows"           "case_when"          
#>  [28] "changes"             "check_dbplyr"        "coalesce"           
#>  [31] "collapse"            "collect"             "combine"            
#>  [34] "common_by"           "compare_tbls"        "compare_tbls2"      
#>  [37] "compute"             "contains"            "copy_to"            
#>  [40] "count"               "count_"              "cumall"             
#>  [43] "cumany"              "cume_dist"           "cummean"            
#>  [46] "current_vars"        "data_frame"          "data_frame_"        
#>  [49] "db_analyze"          "db_begin"            "db_commit"          
#>  [52] "db_create_index"     "db_create_indexes"   "db_create_table"    
#>  [55] "db_data_type"        "db_desc"             "db_drop_table"      
#>  [58] "db_explain"          "db_has_table"        "db_insert_into"     
#>  [61] "db_list_tables"      "db_query_fields"     "db_query_rows"      
#>  [64] "db_rollback"         "db_save_query"       "db_write_table"     
#>  [67] "dense_rank"          "desc"                "dim_desc"           
#>  [70] "distinct"            "distinct_"           "distinct_all"       
#>  [73] "distinct_at"         "distinct_if"         "distinct_prepare"   
#>  [76] "do"                  "do_"                 "dr_dplyr"           
#>  [79] "ends_with"           "enexpr"              "enexprs"            
#>  [82] "enquo"               "enquos"              "ensym"              
#>  [85] "ensyms"              "eval_tbls"           "eval_tbls2"         
#>  [88] "everything"          "explain"             "expr"               
#>  [91] "failwith"            "filter"              "filter_"            
#>  [94] "filter_all"          "filter_at"           "filter_if"          
#>  [97] "first"               "frame_data"          "full_join"          
#> [100] "funs"                "funs_"               "glimpse"            
#> [103] "group_by"            "group_by_"           "group_by_all"       
#> [106] "group_by_at"         "group_by_if"         "group_by_prepare"   
#> [109] "group_cols"          "group_data"          "group_indices"      
#> [112] "group_indices_"      "group_keys"          "group_map"          
#> [115] "group_nest"          "group_rows"          "group_size"         
#> [118] "group_split"         "group_trim"          "group_vars"         
#> [121] "group_walk"          "grouped_df"          "groups"             
#> [124] "hybrid_call"         "id"                  "ident"              
#> [127] "if_else"             "inner_join"          "intersect"          
#> [130] "is.grouped_df"       "is.src"              "is.tbl"             
#> [133] "is_grouped_df"       "lag"                 "last"               
#> [136] "last_col"            "lead"                "left_join"          
#> [139] "location"            "lst"                 "lst_"               
#> [142] "make_tbl"            "matches"             "min_rank"           
#> [145] "mutate"              "mutate_"             "mutate_all"         
#> [148] "mutate_at"           "mutate_each"         "mutate_each_"       
#> [151] "mutate_if"           "n"                   "n_distinct"         
#> [154] "n_groups"            "na_if"               "near"               
#> [157] "nest_join"           "new_grouped_df"      "nth"                
#> [160] "ntile"               "num_range"           "one_of"             
#> [163] "order_by"            "percent_rank"        "progress_estimated" 
#> [166] "pull"                "quo"                 "quo_name"           
#> [169] "quos"                "rbind_all"           "rbind_list"         
#> [172] "recode"              "recode_factor"       "rename"             
#> [175] "rename_"             "rename_all"          "rename_at"          
#> [178] "rename_if"           "rename_vars"         "rename_vars_"       
#> [181] "right_join"          "row_number"          "rowwise"            
#> [184] "same_src"            "sample_frac"         "sample_n"           
#> [187] "select"              "select_"             "select_all"         
#> [190] "select_at"           "select_if"           "select_var"         
#> [193] "select_vars"         "select_vars_"        "semi_join"          
#> [196] "setdiff"             "setequal"            "show_query"         
#> [199] "slice"               "slice_"              "sql"                
#> [202] "sql_escape_ident"    "sql_escape_string"   "sql_join"           
#> [205] "sql_select"          "sql_semi_join"       "sql_set_op"         
#> [208] "sql_subquery"        "sql_translate_env"   "src"                
#> [211] "src_df"              "src_local"           "src_mysql"          
#> [214] "src_postgres"        "src_sqlite"          "src_tbls"           
#> [217] "starts_with"         "summarise"           "summarise_"         
#> [220] "summarise_all"       "summarise_at"        "summarise_each"     
#> [223] "summarise_each_"     "summarise_if"        "summarize"          
#> [226] "summarize_"          "summarize_all"       "summarize_at"       
#> [229] "summarize_each"      "summarize_each_"     "summarize_if"       
#> [232] "sym"                 "syms"                "tally"              
#> [235] "tally_"              "tbl"                 "tbl_cube"           
#> [238] "tbl_df"              "tbl_nongroup_vars"   "tbl_sum"            
#> [241] "tbl_vars"            "tibble"              "top_n"              
#> [244] "transmute"           "transmute_"          "transmute_all"      
#> [247] "transmute_at"        "transmute_if"        "tribble"            
#> [250] "trunc_mat"           "type_sum"            "ungroup"            
#> [253] "union"               "union_all"           "validate_grouped_df"
#> [256] "vars"                "with_order"          "wrap_dbplyr_obj"    
#> 
#> $exportPatterns
#> character(0)
#> 
#> $importClasses
#> list()
#> 
#> $importMethods
#> list()
#> 
#> $exportClasses
#> character(0)
#> 
#> $exportMethods
#> character(0)
#> 
#> $exportClassPatterns
#> character(0)
#> 
#> $dynlibs
#>         
#> "dplyr" 
#> 
#> $S3methods
#>        [,1]              [,2]             [,3] [,4]
#>   [1,] "["               "fun_list"       NA   NA  
#>   [2,] "["               "grouped_df"     NA   NA  
#>   [3,] "all.equal"       "tbl_df"         NA   NA  
#>   [4,] "anti_join"       "data.frame"     NA   NA  
#>   [5,] "anti_join"       "tbl_df"         NA   NA  
#>   [6,] "arrange"         "data.frame"     NA   NA  
#>   [7,] "arrange"         "default"        NA   NA  
#>   [8,] "arrange"         "grouped_df"     NA   NA  
#>   [9,] "arrange"         "tbl_df"         NA   NA  
#>  [10,] "arrange_"        "data.frame"     NA   NA  
#>  [11,] "arrange_"        "tbl_df"         NA   NA  
#>  [12,] "as.data.frame"   "grouped_df"     NA   NA  
#>  [13,] "as.data.frame"   "rowwise_df"     NA   NA  
#>  [14,] "as.data.frame"   "tbl_cube"       NA   NA  
#>  [15,] "as.table"        "tbl_cube"       NA   NA  
#>  [16,] "as.tbl"          "data.frame"     NA   NA  
#>  [17,] "as.tbl"          "tbl"            NA   NA  
#>  [18,] "as.tbl_cube"     "array"          NA   NA  
#>  [19,] "as.tbl_cube"     "data.frame"     NA   NA  
#>  [20,] "as.tbl_cube"     "matrix"         NA   NA  
#>  [21,] "as.tbl_cube"     "table"          NA   NA  
#>  [22,] "as_tibble"       "grouped_df"     NA   NA  
#>  [23,] "as_tibble"       "tbl_cube"       NA   NA  
#>  [24,] "auto_copy"       "tbl_cube"       NA   NA  
#>  [25,] "auto_copy"       "tbl_df"         NA   NA  
#>  [26,] "cbind"           "grouped_df"     NA   NA  
#>  [27,] "collapse"        "data.frame"     NA   NA  
#>  [28,] "collect"         "data.frame"     NA   NA  
#>  [29,] "common_by"       "NULL"           NA   NA  
#>  [30,] "common_by"       "character"      NA   NA  
#>  [31,] "common_by"       "default"        NA   NA  
#>  [32,] "common_by"       "list"           NA   NA  
#>  [33,] "compute"         "data.frame"     NA   NA  
#>  [34,] "copy_to"         "DBIConnection"  NA   NA  
#>  [35,] "copy_to"         "src_local"      NA   NA  
#>  [36,] "default_missing" "data.frame"     NA   NA  
#>  [37,] "default_missing" "default"        NA   NA  
#>  [38,] "dim"             "tbl_cube"       NA   NA  
#>  [39,] "distinct"        "data.frame"     NA   NA  
#>  [40,] "distinct"        "default"        NA   NA  
#>  [41,] "distinct"        "grouped_df"     NA   NA  
#>  [42,] "distinct"        "tbl_df"         NA   NA  
#>  [43,] "distinct_"       "data.frame"     NA   NA  
#>  [44,] "distinct_"       "grouped_df"     NA   NA  
#>  [45,] "distinct_"       "tbl_df"         NA   NA  
#>  [46,] "do"              "NULL"           NA   NA  
#>  [47,] "do"              "data.frame"     NA   NA  
#>  [48,] "do"              "default"        NA   NA  
#>  [49,] "do"              "grouped_df"     NA   NA  
#>  [50,] "do"              "rowwise_df"     NA   NA  
#>  [51,] "do_"             "NULL"           NA   NA  
#>  [52,] "do_"             "data.frame"     NA   NA  
#>  [53,] "do_"             "grouped_df"     NA   NA  
#>  [54,] "do_"             "rowwise_df"     NA   NA  
#>  [55,] "filter"          "data.frame"     NA   NA  
#>  [56,] "filter"          "default"        NA   NA  
#>  [57,] "filter"          "tbl_cube"       NA   NA  
#>  [58,] "filter"          "tbl_df"         NA   NA  
#>  [59,] "filter"          "ts"             NA   NA  
#>  [60,] "filter_"         "data.frame"     NA   NA  
#>  [61,] "filter_"         "tbl_cube"       NA   NA  
#>  [62,] "filter_"         "tbl_df"         NA   NA  
#>  [63,] "format"          "src_local"      NA   NA  
#>  [64,] "full_join"       "data.frame"     NA   NA  
#>  [65,] "full_join"       "tbl_df"         NA   NA  
#>  [66,] "group_by"        "data.frame"     NA   NA  
#>  [67,] "group_by"        "default"        NA   NA  
#>  [68,] "group_by"        "rowwise_df"     NA   NA  
#>  [69,] "group_by"        "tbl_cube"       NA   NA  
#>  [70,] "group_by_"       "data.frame"     NA   NA  
#>  [71,] "group_by_"       "rowwise_df"     NA   NA  
#>  [72,] "group_by_"       "tbl_cube"       NA   NA  
#>  [73,] "group_data"      "data.frame"     NA   NA  
#>  [74,] "group_data"      "grouped_df"     NA   NA  
#>  [75,] "group_data"      "rowwise_df"     NA   NA  
#>  [76,] "group_indices"   "data.frame"     NA   NA  
#>  [77,] "group_indices"   "default"        NA   NA  
#>  [78,] "group_indices"   "grouped_df"     NA   NA  
#>  [79,] "group_indices"   "rowwise_df"     NA   NA  
#>  [80,] "group_indices_"  "data.frame"     NA   NA  
#>  [81,] "group_indices_"  "grouped_df"     NA   NA  
#>  [82,] "group_indices_"  "rowwise_df"     NA   NA  
#>  [83,] "group_keys"      "data.frame"     NA   NA  
#>  [84,] "group_keys"      "grouped_df"     NA   NA  
#>  [85,] "group_keys"      "rowwise_df"     NA   NA  
#>  [86,] "group_map"       "function"       NA   NA  
#>  [87,] "group_map"       "data.frame"     NA   NA  
#>  [88,] "group_map"       "formula"        NA   NA  
#>  [89,] "group_map"       "grouped_df"     NA   NA  
#>  [90,] "group_nest"      "data.frame"     NA   NA  
#>  [91,] "group_nest"      "grouped_df"     NA   NA  
#>  [92,] "group_size"      "data.frame"     NA   NA  
#>  [93,] "group_size"      "grouped_df"     NA   NA  
#>  [94,] "group_size"      "rowwise_df"     NA   NA  
#>  [95,] "group_split"     "data.frame"     NA   NA  
#>  [96,] "group_split"     "grouped_df"     NA   NA  
#>  [97,] "group_split"     "rowwise_df"     NA   NA  
#>  [98,] "group_trim"      "data.frame"     NA   NA  
#>  [99,] "group_trim"      "grouped_df"     NA   NA  
#> [100,] "group_vars"      "default"        NA   NA  
#> [101,] "group_vars"      "grouped_df"     NA   NA  
#> [102,] "group_vars"      "tbl_cube"       NA   NA  
#> [103,] "group_walk"      "data.frame"     NA   NA  
#> [104,] "group_walk"      "grouped_df"     NA   NA  
#> [105,] "groups"          "data.frame"     NA   NA  
#> [106,] "groups"          "grouped_df"     NA   NA  
#> [107,] "groups"          "tbl_cube"       NA   NA  
#> [108,] "hybrid_call"     "data.frame"     NA   NA  
#> [109,] "inner_join"      "data.frame"     NA   NA  
#> [110,] "inner_join"      "tbl_df"         NA   NA  
#> [111,] "intersect"       "data.frame"     NA   NA  
#> [112,] "intersect"       "default"        NA   NA  
#> [113,] "left_join"       "data.frame"     NA   NA  
#> [114,] "left_join"       "tbl_df"         NA   NA  
#> [115,] "mutate"          "data.frame"     NA   NA  
#> [116,] "mutate"          "default"        NA   NA  
#> [117,] "mutate"          "tbl_df"         NA   NA  
#> [118,] "mutate_"         "data.frame"     NA   NA  
#> [119,] "mutate_"         "tbl_df"         NA   NA  
#> [120,] "n_groups"        "data.frame"     NA   NA  
#> [121,] "n_groups"        "grouped_df"     NA   NA  
#> [122,] "n_groups"        "rowwise_df"     NA   NA  
#> [123,] "nest_join"       "data.frame"     NA   NA  
#> [124,] "nest_join"       "tbl_df"         NA   NA  
#> [125,] "print"           "BoolResult"     NA   NA  
#> [126,] "print"           "all_vars"       NA   NA  
#> [127,] "print"           "any_vars"       NA   NA  
#> [128,] "print"           "dplyr_sel_vars" NA   NA  
#> [129,] "print"           "fun_list"       NA   NA  
#> [130,] "print"           "hybrid_call"    NA   NA  
#> [131,] "print"           "location"       NA   NA  
#> [132,] "print"           "rowwise_df"     NA   NA  
#> [133,] "print"           "src"            NA   NA  
#> [134,] "print"           "tbl_cube"       NA   NA  
#> [135,] "pull"            "data.frame"     NA   NA  
#> [136,] "rbind"           "grouped_df"     NA   NA  
#> [137,] "recode"          "character"      NA   NA  
#> [138,] "recode"          "factor"         NA   NA  
#> [139,] "recode"          "numeric"        NA   NA  
#> [140,] "rename"          "data.frame"     NA   NA  
#> [141,] "rename"          "default"        NA   NA  
#> [142,] "rename"          "grouped_df"     NA   NA  
#> [143,] "rename"          "tbl_cube"       NA   NA  
#> [144,] "rename_"         "data.frame"     NA   NA  
#> [145,] "rename_"         "grouped_df"     NA   NA  
#> [146,] "rename_"         "tbl_cube"       NA   NA  
#> [147,] "right_join"      "data.frame"     NA   NA  
#> [148,] "right_join"      "tbl_df"         NA   NA  
#> [149,] "same_src"        "data.frame"     NA   NA  
#> [150,] "same_src"        "tbl_cube"       NA   NA  
#> [151,] "sample_frac"     "data.frame"     NA   NA  
#> [152,] "sample_frac"     "default"        NA   NA  
#> [153,] "sample_n"        "data.frame"     NA   NA  
#> [154,] "sample_n"        "default"        NA   NA  
#> [155,] "select"          "data.frame"     NA   NA  
#> [156,] "select"          "default"        NA   NA  
#> [157,] "select"          "grouped_df"     NA   NA  
#> [158,] "select"          "tbl_cube"       NA   NA  
#> [159,] "select_"         "data.frame"     NA   NA  
#> [160,] "select_"         "grouped_df"     NA   NA  
#> [161,] "select_"         "tbl_cube"       NA   NA  
#> [162,] "semi_join"       "data.frame"     NA   NA  
#> [163,] "semi_join"       "tbl_df"         NA   NA  
#> [164,] "setdiff"         "data.frame"     NA   NA  
#> [165,] "setdiff"         "default"        NA   NA  
#> [166,] "setequal"        "data.frame"     NA   NA  
#> [167,] "setequal"        "default"        NA   NA  
#> [168,] "slice"           "data.frame"     NA   NA  
#> [169,] "slice"           "default"        NA   NA  
#> [170,] "slice"           "tbl_df"         NA   NA  
#> [171,] "slice_"          "data.frame"     NA   NA  
#> [172,] "slice_"          "tbl_df"         NA   NA  
#> [173,] "src_tbls"        "src_local"      NA   NA  
#> [174,] "summarise"       "data.frame"     NA   NA  
#> [175,] "summarise"       "default"        NA   NA  
#> [176,] "summarise"       "tbl_cube"       NA   NA  
#> [177,] "summarise"       "tbl_df"         NA   NA  
#> [178,] "summarise_"      "data.frame"     NA   NA  
#> [179,] "summarise_"      "tbl_cube"       NA   NA  
#> [180,] "summarise_"      "tbl_df"         NA   NA  
#> [181,] "tbl"             "DBIConnection"  NA   NA  
#> [182,] "tbl"             "src_local"      NA   NA  
#> [183,] "tbl_sum"         "grouped_df"     NA   NA  
#> [184,] "tbl_vars"        "data.frame"     NA   NA  
#> [185,] "tbl_vars"        "tbl_cube"       NA   NA  
#> [186,] "transmute"       "default"        NA   NA  
#> [187,] "transmute"       "grouped_df"     NA   NA  
#> [188,] "transmute_"      "default"        NA   NA  
#> [189,] "transmute_"      "grouped_df"     NA   NA  
#> [190,] "ungroup"         "data.frame"     NA   NA  
#> [191,] "ungroup"         "grouped_df"     NA   NA  
#> [192,] "ungroup"         "rowwise_df"     NA   NA  
#> [193,] "union"           "data.frame"     NA   NA  
#> [194,] "union"           "default"        NA   NA  
#> [195,] "union_all"       "data.frame"     NA   NA  
#> [196,] "union_all"       "default"        NA   NA  
#> 
pac_namespace("dplyr", at = as.Date("2019-02-01"))
#> $imports
#> $imports[[1]]
#> [1] "rlang"
#> 
#> $imports[[2]]
#> $imports[[2]][[1]]
#> [1] "R6"
#> 
#> $imports[[2]][[2]]
#> [1] "R6Class"
#> 
#> 
#> $imports[[3]]
#> $imports[[3]][[1]]
#> [1] "Rcpp"
#> 
#> $imports[[3]][[2]]
#> [1] "Rcpp.plugin.maker"
#> 
#> 
#> $imports[[4]]
#> $imports[[4]][[1]]
#> [1] "Rcpp"
#> 
#> $imports[[4]][[2]]
#> [1] "cppFunction"
#> 
#> 
#> $imports[[5]]
#> $imports[[5]][[1]]
#> [1] "assertthat"
#> 
#> $imports[[5]][[2]]
#> [1] "on_failure<-"
#> 
#> 
#> $imports[[6]]
#> $imports[[6]][[1]]
#> [1] "assertthat"
#> 
#> $imports[[6]][[2]]
#> [1] "assert_that"
#> 
#> 
#> $imports[[7]]
#> $imports[[7]][[1]]
#> [1] "assertthat"
#> 
#> $imports[[7]][[2]]
#> [1] "is.flag"
#> 
#> 
#> $imports[[8]]
#> $imports[[8]][[1]]
#> [1] "bindrcpp"
#> 
#> $imports[[8]][[2]]
#> [1] "create_env"
#> 
#> 
#> $imports[[9]]
#> $imports[[9]][[1]]
#> [1] "glue"
#> 
#> $imports[[9]][[2]]
#> [1] "glue"
#> 
#> 
#> $imports[[10]]
#> $imports[[10]][[1]]
#> [1] "magrittr"
#> 
#> $imports[[10]][[2]]
#> [1] "%>%"
#> 
#> 
#> $imports[[11]]
#> $imports[[11]][[1]]
#> [1] "methods"
#> 
#> $imports[[11]][[2]]
#> [1] "is"
#> 
#> 
#> $imports[[12]]
#> $imports[[12]][[1]]
#> [1] "pkgconfig"
#> 
#> $imports[[12]][[2]]
#> [1] "get_config"
#> 
#> 
#> $imports[[13]]
#> $imports[[13]][[1]]
#> [1] "stats"
#> 
#> $imports[[13]][[2]]
#> [1] "lag"
#> 
#> 
#> $imports[[14]]
#> $imports[[14]][[1]]
#> [1] "stats"
#> 
#> $imports[[14]][[2]]
#> [1] "setNames"
#> 
#> 
#> $imports[[15]]
#> $imports[[15]][[1]]
#> [1] "stats"
#> 
#> $imports[[15]][[2]]
#> [1] "update"
#> 
#> 
#> $imports[[16]]
#> $imports[[16]][[1]]
#> [1] "tibble"
#> 
#> $imports[[16]][[2]]
#> [1] "add_row"
#> 
#> 
#> $imports[[17]]
#> $imports[[17]][[1]]
#> [1] "tibble"
#> 
#> $imports[[17]][[2]]
#> [1] "as_data_frame"
#> 
#> 
#> $imports[[18]]
#> $imports[[18]][[1]]
#> [1] "tibble"
#> 
#> $imports[[18]][[2]]
#> [1] "as_tibble"
#> 
#> 
#> $imports[[19]]
#> $imports[[19]][[1]]
#> [1] "tibble"
#> 
#> $imports[[19]][[2]]
#> [1] "data_frame"
#> 
#> 
#> $imports[[20]]
#> $imports[[20]][[1]]
#> [1] "tibble"
#> 
#> $imports[[20]][[2]]
#> [1] "data_frame_"
#> 
#> 
#> $imports[[21]]
#> $imports[[21]][[1]]
#> [1] "tibble"
#> 
#> $imports[[21]][[2]]
#> [1] "frame_data"
#> 
#> 
#> $imports[[22]]
#> $imports[[22]][[1]]
#> [1] "tibble"
#> 
#> $imports[[22]][[2]]
#> [1] "glimpse"
#> 
#> 
#> $imports[[23]]
#> $imports[[23]][[1]]
#> [1] "tibble"
#> 
#> $imports[[23]][[2]]
#> [1] "lst"
#> 
#> 
#> $imports[[24]]
#> $imports[[24]][[1]]
#> [1] "tibble"
#> 
#> $imports[[24]][[2]]
#> [1] "lst_"
#> 
#> 
#> $imports[[25]]
#> $imports[[25]][[1]]
#> [1] "tibble"
#> 
#> $imports[[25]][[2]]
#> [1] "tbl_sum"
#> 
#> 
#> $imports[[26]]
#> $imports[[26]][[1]]
#> [1] "tibble"
#> 
#> $imports[[26]][[2]]
#> [1] "tibble"
#> 
#> 
#> $imports[[27]]
#> $imports[[27]][[1]]
#> [1] "tibble"
#> 
#> $imports[[27]][[2]]
#> [1] "tribble"
#> 
#> 
#> $imports[[28]]
#> $imports[[28]][[1]]
#> [1] "tibble"
#> 
#> $imports[[28]][[2]]
#> [1] "trunc_mat"
#> 
#> 
#> $imports[[29]]
#> $imports[[29]][[1]]
#> [1] "tibble"
#> 
#> $imports[[29]][[2]]
#> [1] "type_sum"
#> 
#> 
#> $imports[[30]]
#> $imports[[30]][[1]]
#> [1] "tidyselect"
#> 
#> $imports[[30]][[2]]
#> [1] "contains"
#> 
#> 
#> $imports[[31]]
#> $imports[[31]][[1]]
#> [1] "tidyselect"
#> 
#> $imports[[31]][[2]]
#> [1] "ends_with"
#> 
#> 
#> $imports[[32]]
#> $imports[[32]][[1]]
#> [1] "tidyselect"
#> 
#> $imports[[32]][[2]]
#> [1] "everything"
#> 
#> 
#> $imports[[33]]
#> $imports[[33]][[1]]
#> [1] "tidyselect"
#> 
#> $imports[[33]][[2]]
#> [1] "matches"
#> 
#> 
#> $imports[[34]]
#> $imports[[34]][[1]]
#> [1] "tidyselect"
#> 
#> $imports[[34]][[2]]
#> [1] "num_range"
#> 
#> 
#> $imports[[35]]
#> $imports[[35]][[1]]
#> [1] "tidyselect"
#> 
#> $imports[[35]][[2]]
#> [1] "one_of"
#> 
#> 
#> $imports[[36]]
#> $imports[[36]][[1]]
#> [1] "tidyselect"
#> 
#> $imports[[36]][[2]]
#> [1] "starts_with"
#> 
#> 
#> $imports[[37]]
#> $imports[[37]][[1]]
#> [1] "utils"
#> 
#> $imports[[37]][[2]]
#> [1] "head"
#> 
#> 
#> $imports[[38]]
#> $imports[[38]][[1]]
#> [1] "utils"
#> 
#> $imports[[38]][[2]]
#> [1] "tail"
#> 
#> 
#> 
#> $exports
#>   [1] "%>%"                ".data"              "add_count"         
#>   [4] "add_count_"         "add_row"            "add_rownames"      
#>   [7] "add_tally"          "add_tally_"         "all_equal"         
#>  [10] "all_vars"           "anti_join"          "any_vars"          
#>  [13] "arrange"            "arrange_"           "arrange_all"       
#>  [16] "arrange_at"         "arrange_if"         "as.tbl"            
#>  [19] "as.tbl_cube"        "as_data_frame"      "as_tibble"         
#>  [22] "auto_copy"          "bench_tbls"         "between"           
#>  [25] "bind_cols"          "bind_rows"          "case_when"         
#>  [28] "changes"            "check_dbplyr"       "coalesce"          
#>  [31] "collapse"           "collect"            "combine"           
#>  [34] "common_by"          "compare_tbls"       "compare_tbls2"     
#>  [37] "compute"            "contains"           "copy_to"           
#>  [40] "count"              "count_"             "cumall"            
#>  [43] "cumany"             "cume_dist"          "cummean"           
#>  [46] "current_vars"       "data_frame"         "data_frame_"       
#>  [49] "db_analyze"         "db_begin"           "db_commit"         
#>  [52] "db_create_index"    "db_create_indexes"  "db_create_table"   
#>  [55] "db_data_type"       "db_desc"            "db_drop_table"     
#>  [58] "db_explain"         "db_has_table"       "db_insert_into"    
#>  [61] "db_list_tables"     "db_query_fields"    "db_query_rows"     
#>  [64] "db_rollback"        "db_save_query"      "db_write_table"    
#>  [67] "dense_rank"         "desc"               "dim_desc"          
#>  [70] "distinct"           "distinct_"          "do"                
#>  [73] "do_"                "dr_dplyr"           "ends_with"         
#>  [76] "enexpr"             "enexprs"            "enquo"             
#>  [79] "enquos"             "ensym"              "ensyms"            
#>  [82] "eval_tbls"          "eval_tbls2"         "everything"        
#>  [85] "explain"            "expr"               "failwith"          
#>  [88] "filter"             "filter_"            "filter_all"        
#>  [91] "filter_at"          "filter_if"          "first"             
#>  [94] "frame_data"         "full_join"          "funs"              
#>  [97] "funs_"              "glimpse"            "group_by"          
#> [100] "group_by_"          "group_by_all"       "group_by_at"       
#> [103] "group_by_if"        "group_by_prepare"   "group_indices"     
#> [106] "group_indices_"     "group_size"         "group_vars"        
#> [109] "grouped_df"         "groups"             "id"                
#> [112] "ident"              "if_else"            "inner_join"        
#> [115] "intersect"          "is.grouped_df"      "is.src"            
#> [118] "is.tbl"             "is_grouped_df"      "lag"               
#> [121] "last"               "lead"               "left_join"         
#> [124] "location"           "lst"                "lst_"              
#> [127] "make_tbl"           "matches"            "min_rank"          
#> [130] "mutate"             "mutate_"            "mutate_all"        
#> [133] "mutate_at"          "mutate_each"        "mutate_each_"      
#> [136] "mutate_if"          "n"                  "n_distinct"        
#> [139] "n_groups"           "na_if"              "near"              
#> [142] "nth"                "ntile"              "num_range"         
#> [145] "one_of"             "order_by"           "percent_rank"      
#> [148] "progress_estimated" "pull"               "quo"               
#> [151] "quo_name"           "quos"               "rbind_all"         
#> [154] "rbind_list"         "recode"             "recode_factor"     
#> [157] "rename"             "rename_"            "rename_all"        
#> [160] "rename_at"          "rename_if"          "rename_vars"       
#> [163] "rename_vars_"       "right_join"         "row_number"        
#> [166] "rowwise"            "same_src"           "sample_frac"       
#> [169] "sample_n"           "select"             "select_"           
#> [172] "select_all"         "select_at"          "select_if"         
#> [175] "select_var"         "select_vars"        "select_vars_"      
#> [178] "semi_join"          "setdiff"            "setequal"          
#> [181] "show_query"         "slice"              "slice_"            
#> [184] "sql"                "sql_escape_ident"   "sql_escape_string" 
#> [187] "sql_join"           "sql_select"         "sql_semi_join"     
#> [190] "sql_set_op"         "sql_subquery"       "sql_translate_env" 
#> [193] "src"                "src_df"             "src_local"         
#> [196] "src_mysql"          "src_postgres"       "src_sqlite"        
#> [199] "src_tbls"           "starts_with"        "summarise"         
#> [202] "summarise_"         "summarise_all"      "summarise_at"      
#> [205] "summarise_each"     "summarise_each_"    "summarise_if"      
#> [208] "summarize"          "summarize_"         "summarize_all"     
#> [211] "summarize_at"       "summarize_each"     "summarize_each_"   
#> [214] "summarize_if"       "sym"                "syms"              
#> [217] "tally"              "tally_"             "tbl"               
#> [220] "tbl_cube"           "tbl_df"             "tbl_nongroup_vars" 
#> [223] "tbl_sum"            "tbl_vars"           "tibble"            
#> [226] "top_n"              "transmute"          "transmute_"        
#> [229] "transmute_all"      "transmute_at"       "transmute_if"      
#> [232] "tribble"            "trunc_mat"          "type_sum"          
#> [235] "ungroup"            "union"              "union_all"         
#> [238] "vars"               "with_order"         "wrap_dbplyr_obj"   
#> 
#> $exportPatterns
#> character(0)
#> 
#> $importClasses
#> list()
#> 
#> $importMethods
#> list()
#> 
#> $exportClasses
#> character(0)
#> 
#> $exportMethods
#> character(0)
#> 
#> $exportClassPatterns
#> character(0)
#> 
#> $dynlibs
#>         
#> "dplyr" 
#> 
#> $S3methods
#>        [,1]              [,2]            [,3] [,4]
#>   [1,] "["               "fun_list"      NA   NA  
#>   [2,] "["               "grouped_df"    NA   NA  
#>   [3,] "all.equal"       "tbl_df"        NA   NA  
#>   [4,] "anti_join"       "data.frame"    NA   NA  
#>   [5,] "anti_join"       "tbl_df"        NA   NA  
#>   [6,] "arrange"         "data.frame"    NA   NA  
#>   [7,] "arrange"         "default"       NA   NA  
#>   [8,] "arrange"         "grouped_df"    NA   NA  
#>   [9,] "arrange"         "tbl_df"        NA   NA  
#>  [10,] "arrange_"        "data.frame"    NA   NA  
#>  [11,] "arrange_"        "tbl_df"        NA   NA  
#>  [12,] "as.data.frame"   "grouped_df"    NA   NA  
#>  [13,] "as.data.frame"   "rowwise_df"    NA   NA  
#>  [14,] "as.data.frame"   "tbl_cube"      NA   NA  
#>  [15,] "as.table"        "tbl_cube"      NA   NA  
#>  [16,] "as.tbl"          "data.frame"    NA   NA  
#>  [17,] "as.tbl"          "tbl"           NA   NA  
#>  [18,] "as.tbl_cube"     "array"         NA   NA  
#>  [19,] "as.tbl_cube"     "data.frame"    NA   NA  
#>  [20,] "as.tbl_cube"     "matrix"        NA   NA  
#>  [21,] "as.tbl_cube"     "table"         NA   NA  
#>  [22,] "as_tibble"       "grouped_df"    NA   NA  
#>  [23,] "as_tibble"       "tbl_cube"      NA   NA  
#>  [24,] "auto_copy"       "tbl_cube"      NA   NA  
#>  [25,] "auto_copy"       "tbl_df"        NA   NA  
#>  [26,] "cbind"           "grouped_df"    NA   NA  
#>  [27,] "collapse"        "data.frame"    NA   NA  
#>  [28,] "collect"         "data.frame"    NA   NA  
#>  [29,] "common_by"       "NULL"          NA   NA  
#>  [30,] "common_by"       "character"     NA   NA  
#>  [31,] "common_by"       "default"       NA   NA  
#>  [32,] "common_by"       "list"          NA   NA  
#>  [33,] "compute"         "data.frame"    NA   NA  
#>  [34,] "copy_to"         "DBIConnection" NA   NA  
#>  [35,] "copy_to"         "src_local"     NA   NA  
#>  [36,] "default_missing" "data.frame"    NA   NA  
#>  [37,] "default_missing" "default"       NA   NA  
#>  [38,] "dim"             "tbl_cube"      NA   NA  
#>  [39,] "distinct"        "data.frame"    NA   NA  
#>  [40,] "distinct"        "default"       NA   NA  
#>  [41,] "distinct"        "grouped_df"    NA   NA  
#>  [42,] "distinct"        "tbl_df"        NA   NA  
#>  [43,] "distinct_"       "data.frame"    NA   NA  
#>  [44,] "distinct_"       "grouped_df"    NA   NA  
#>  [45,] "distinct_"       "tbl_df"        NA   NA  
#>  [46,] "do"              "NULL"          NA   NA  
#>  [47,] "do"              "data.frame"    NA   NA  
#>  [48,] "do"              "default"       NA   NA  
#>  [49,] "do"              "grouped_df"    NA   NA  
#>  [50,] "do"              "rowwise_df"    NA   NA  
#>  [51,] "do_"             "NULL"          NA   NA  
#>  [52,] "do_"             "data.frame"    NA   NA  
#>  [53,] "do_"             "grouped_df"    NA   NA  
#>  [54,] "do_"             "rowwise_df"    NA   NA  
#>  [55,] "filter"          "data.frame"    NA   NA  
#>  [56,] "filter"          "default"       NA   NA  
#>  [57,] "filter"          "tbl_cube"      NA   NA  
#>  [58,] "filter"          "tbl_df"        NA   NA  
#>  [59,] "filter"          "ts"            NA   NA  
#>  [60,] "filter_"         "data.frame"    NA   NA  
#>  [61,] "filter_"         "tbl_cube"      NA   NA  
#>  [62,] "filter_"         "tbl_df"        NA   NA  
#>  [63,] "format"          "src_local"     NA   NA  
#>  [64,] "full_join"       "data.frame"    NA   NA  
#>  [65,] "full_join"       "tbl_df"        NA   NA  
#>  [66,] "group_by"        "data.frame"    NA   NA  
#>  [67,] "group_by"        "default"       NA   NA  
#>  [68,] "group_by"        "rowwise_df"    NA   NA  
#>  [69,] "group_by"        "tbl_cube"      NA   NA  
#>  [70,] "group_by_"       "data.frame"    NA   NA  
#>  [71,] "group_by_"       "rowwise_df"    NA   NA  
#>  [72,] "group_by_"       "tbl_cube"      NA   NA  
#>  [73,] "group_indices"   "data.frame"    NA   NA  
#>  [74,] "group_indices"   "default"       NA   NA  
#>  [75,] "group_indices"   "grouped_df"    NA   NA  
#>  [76,] "group_indices_"  "data.frame"    NA   NA  
#>  [77,] "group_indices_"  "grouped_df"    NA   NA  
#>  [78,] "group_size"      "data.frame"    NA   NA  
#>  [79,] "group_size"      "grouped_df"    NA   NA  
#>  [80,] "group_size"      "rowwise_df"    NA   NA  
#>  [81,] "group_vars"      "default"       NA   NA  
#>  [82,] "group_vars"      "grouped_df"    NA   NA  
#>  [83,] "group_vars"      "tbl_cube"      NA   NA  
#>  [84,] "groups"          "data.frame"    NA   NA  
#>  [85,] "groups"          "grouped_df"    NA   NA  
#>  [86,] "groups"          "tbl_cube"      NA   NA  
#>  [87,] "inner_join"      "data.frame"    NA   NA  
#>  [88,] "inner_join"      "tbl_df"        NA   NA  
#>  [89,] "intersect"       "data.frame"    NA   NA  
#>  [90,] "intersect"       "default"       NA   NA  
#>  [91,] "left_join"       "data.frame"    NA   NA  
#>  [92,] "left_join"       "tbl_df"        NA   NA  
#>  [93,] "mutate"          "data.frame"    NA   NA  
#>  [94,] "mutate"          "default"       NA   NA  
#>  [95,] "mutate"          "tbl_df"        NA   NA  
#>  [96,] "mutate_"         "data.frame"    NA   NA  
#>  [97,] "mutate_"         "tbl_df"        NA   NA  
#>  [98,] "n_groups"        "data.frame"    NA   NA  
#>  [99,] "n_groups"        "grouped_df"    NA   NA  
#> [100,] "n_groups"        "rowwise_df"    NA   NA  
#> [101,] "print"           "BoolResult"    NA   NA  
#> [102,] "print"           "all_vars"      NA   NA  
#> [103,] "print"           "any_vars"      NA   NA  
#> [104,] "print"           "fun_list"      NA   NA  
#> [105,] "print"           "location"      NA   NA  
#> [106,] "print"           "rowwise_df"    NA   NA  
#> [107,] "print"           "src"           NA   NA  
#> [108,] "print"           "tbl_cube"      NA   NA  
#> [109,] "pull"            "data.frame"    NA   NA  
#> [110,] "rbind"           "grouped_df"    NA   NA  
#> [111,] "recode"          "character"     NA   NA  
#> [112,] "recode"          "factor"        NA   NA  
#> [113,] "recode"          "numeric"       NA   NA  
#> [114,] "rename"          "data.frame"    NA   NA  
#> [115,] "rename"          "default"       NA   NA  
#> [116,] "rename"          "grouped_df"    NA   NA  
#> [117,] "rename"          "tbl_cube"      NA   NA  
#> [118,] "rename_"         "data.frame"    NA   NA  
#> [119,] "rename_"         "grouped_df"    NA   NA  
#> [120,] "rename_"         "tbl_cube"      NA   NA  
#> [121,] "right_join"      "data.frame"    NA   NA  
#> [122,] "right_join"      "tbl_df"        NA   NA  
#> [123,] "same_src"        "data.frame"    NA   NA  
#> [124,] "same_src"        "tbl_cube"      NA   NA  
#> [125,] "sample_frac"     "data.frame"    NA   NA  
#> [126,] "sample_frac"     "default"       NA   NA  
#> [127,] "sample_frac"     "grouped_df"    NA   NA  
#> [128,] "sample_n"        "data.frame"    NA   NA  
#> [129,] "sample_n"        "default"       NA   NA  
#> [130,] "sample_n"        "grouped_df"    NA   NA  
#> [131,] "select"          "data.frame"    NA   NA  
#> [132,] "select"          "default"       NA   NA  
#> [133,] "select"          "grouped_df"    NA   NA  
#> [134,] "select"          "tbl_cube"      NA   NA  
#> [135,] "select_"         "data.frame"    NA   NA  
#> [136,] "select_"         "grouped_df"    NA   NA  
#> [137,] "select_"         "tbl_cube"      NA   NA  
#> [138,] "semi_join"       "data.frame"    NA   NA  
#> [139,] "semi_join"       "tbl_df"        NA   NA  
#> [140,] "setdiff"         "data.frame"    NA   NA  
#> [141,] "setdiff"         "default"       NA   NA  
#> [142,] "setequal"        "data.frame"    NA   NA  
#> [143,] "setequal"        "default"       NA   NA  
#> [144,] "slice"           "data.frame"    NA   NA  
#> [145,] "slice"           "default"       NA   NA  
#> [146,] "slice"           "tbl_df"        NA   NA  
#> [147,] "slice_"          "data.frame"    NA   NA  
#> [148,] "slice_"          "tbl_df"        NA   NA  
#> [149,] "src_tbls"        "src_local"     NA   NA  
#> [150,] "summarise"       "data.frame"    NA   NA  
#> [151,] "summarise"       "default"       NA   NA  
#> [152,] "summarise"       "tbl_cube"      NA   NA  
#> [153,] "summarise"       "tbl_df"        NA   NA  
#> [154,] "summarise_"      "data.frame"    NA   NA  
#> [155,] "summarise_"      "tbl_cube"      NA   NA  
#> [156,] "summarise_"      "tbl_df"        NA   NA  
#> [157,] "tbl"             "DBIConnection" NA   NA  
#> [158,] "tbl"             "src_local"     NA   NA  
#> [159,] "tbl_sum"         "grouped_df"    NA   NA  
#> [160,] "tbl_vars"        "data.frame"    NA   NA  
#> [161,] "tbl_vars"        "tbl_cube"      NA   NA  
#> [162,] "transmute"       "default"       NA   NA  
#> [163,] "transmute"       "grouped_df"    NA   NA  
#> [164,] "transmute_"      "default"       NA   NA  
#> [165,] "transmute_"      "grouped_df"    NA   NA  
#> [166,] "ungroup"         "data.frame"    NA   NA  
#> [167,] "ungroup"         "grouped_df"    NA   NA  
#> [168,] "ungroup"         "rowwise_df"    NA   NA  
#> [169,] "union"           "data.frame"    NA   NA  
#> [170,] "union"           "default"       NA   NA  
#> [171,] "union_all"       "data.frame"    NA   NA  
#> [172,] "union_all"       "default"       NA   NA  
#> 
pac_namespace("memoise", local = TRUE)
#> $imports
#> $imports[[1]]
#> $imports[[1]][[1]]
#> [1] "cachem"
#> 
#> $imports[[1]][[2]]
#> [1] "key_missing"
#> 
#> 
#> $imports[[2]]
#> $imports[[2]][[1]]
#> [1] "stats"
#> 
#> $imports[[2]][[2]]
#> [1] "setNames"
#> 
#> 
#> 
#> $exports
#>  [1] "cache_filesystem" "cache_gcs"        "cache_memory"     "cache_s3"        
#>  [5] "drop_cache"       "forget"           "has_cache"        "is.memoised"     
#>  [9] "is.memoized"      "memoise"          "memoize"          "timeout"         
#> 
#> $exportPatterns
#> character(0)
#> 
#> $importClasses
#> list()
#> 
#> $importMethods
#> list()
#> 
#> $exportClasses
#> character(0)
#> 
#> $exportMethods
#> character(0)
#> 
#> $exportClassPatterns
#> character(0)
#> 
#> $dynlibs
#> character(0)
#> 
#> $S3methods
#>      [,1]    [,2]       [,3] [,4]
#> [1,] "print" "memoised" NA   NA  
#>