Package version at a specific Date or a Date interval

using CRAN website to get a package version/versions used at a specific Date or a Date interval.

pac_timemachine(
  pac,
  at = NULL,
  from = NULL,
  to = NULL,
  version = NULL,
  repos = "https://cran.rstudio.com/"
)

Arguments

pac
character a package name.

at
Date old version of package. Default: NULL

from
Date new version of package. Default: NULL

to
Date CRAN URL. Default: NULL

version
character version of package. Default: NULL

repos
character the base URL of the repositories to use. Default https://cran.rstudio.com/

Value

data.frame

Note

Function will scrap two CRAN URLS. Works only with CRAN packages. Please as a courtesy to the R CRAN, don't overload their servers by constantly using this function. The base part of URL in the result is https://cran.r-project.org/src/contrib/. Results are cached for 1 hour with memoise package.

Examples

pac_timemachine("dplyr", at = as.Date("2017-02-02"))
#>    Package Version   Released   Archived Life_Duration
#> 13   dplyr   0.5.0 2016-06-24 2017-06-09      350 days
#>                                 URL Size
#> 13 Archive/dplyr/dplyr_0.5.0.tar.gz 692K
pac_timemachine("dplyr", from = as.Date("2017-02-02"), to = as.Date("2018-04-02"))
#>    Package Version   Released   Archived Life_Duration
#> 13   dplyr   0.5.0 2016-06-24 2017-06-09      350 days
#> 14   dplyr   0.7.0 2017-06-09 2017-06-22       13 days
#> 15   dplyr   0.7.1 2017-06-22 2017-07-21       29 days
#> 16   dplyr   0.7.2 2017-07-21 2017-09-09       50 days
#> 17   dplyr   0.7.3 2017-09-09 2017-09-28       19 days
#> 18   dplyr   0.7.4 2017-09-28 2018-05-19      233 days
#>                                 URL Size
#> 13 Archive/dplyr/dplyr_0.5.0.tar.gz 692K
#> 14 Archive/dplyr/dplyr_0.7.0.tar.gz 675K
#> 15 Archive/dplyr/dplyr_0.7.1.tar.gz 676K
#> 16 Archive/dplyr/dplyr_0.7.2.tar.gz 677K
#> 17 Archive/dplyr/dplyr_0.7.3.tar.gz 789K
#> 18 Archive/dplyr/dplyr_0.7.4.tar.gz 789K
pac_timemachine("dplyr", at = Sys.Date())
#>     Package Version   Released Archived Life_Duration                URL Size
#> 110   dplyr   1.0.7 2021-06-18     <NA>       48 days dplyr_1.0.7.tar.gz <NA>