Validate a specific local package

Checking if installed package dependencies have correct versions taking into account their DESCRIPTION files requirements. Moreover identifying which packages are newest releases. Optionally we could add life duration and CRAN check page status for each dependency.

pac_validate(
  pac,
  lib.loc = NULL,
  fields = c("Depends", "Imports", "LinkingTo"),
  lifeduration = FALSE,
  checkred = list(scope = character(0), flavors = NULL),
  repos = biocran_repos()
)

Arguments

pac
character a package name.

lib.loc
character. Default: NULL

fields
character vector with possible values c("Depends", "Imports", "LinkingTo", "Suggests"). Default: c("Depends", "Imports", "LinkingTo")

lifeduration
logical if to add life duration column, might take some time. Default: FALSE

checkred
list with two named fields, scope and flavor. scope of R CRAN check pages statuses to consider, any of c("ERROR", "FAIL", "WARN", "NOTE"). flavor vector of machines to consider, which might be retrieved with pacs::cran_flavors()$Flavor. By default an empty scope field deactivated assessment for checkred column, and NULL flavor will results in checking all machines. Default list(scope = character(0), flavor = NULL)

repos
character vector base URLs of the repositories to use. By default checking CRAN and newest Bioconductor per R version. Default pacs::biocran_repos()

Value

data.frame with 5/6/7 columns.

Package

character a package name.

Version.expected.min

character expected by DESCRIPTION files minimal version. "" means not specified so the newest version.

Version.have

character installed package version.

version_status

numeric -1/0/1 which comes from utils::compareVersion function. 0 means that we have the same version as required by DESCRIPTION files. -1 means we have too low version installed, this is an error. 1 means we have higher version.

direct

logical if the package is in the first depencency layer, direct depencencies from DESCRIPTION file.

newest

logical if the installed version is the newest one.

cran

logical if the package is on CRAN, version is not taken into account here.

checkred

(Optional) logical if the NEWEST package contains any specified statuses on CRAN check page.

lifeduration

(Optional) integer number of days a package was released.

Note

Version.expected.min column not count packages which are not a dependency for any package, so could not be find in DESCRIPTION files. When turn on the lifeduration option, calculations might be time consuming. Please as a courtesy to the R CRAN, don't overload their server by constantly using this function with lifeduration or checkred turned on. Results are cached with memoise package, memory cache.

Examples

if (FALSE) {
pac_validate("memoise")
pac_validate("memoise",
  lifeduration = TRUE,
  checkred = list(scope = c("ERROR", "FAIL"), flavors = NULL)
)
}