packages DESCRIPTION files

packages DESCRIPTION files taken locally or if remotely then from GITHUB CRAN mirror or CRAN website.

pacs_description(
  pacs,
  versions = NULL,
  at = NULL,
  local = FALSE,
  lib.loc = NULL
)

Arguments

pacs
character vector packages names.

versions
character vector versions. Default: NULL

at
Date . Default: NULL

local
logical if to use local library. Default: FALSE

lib.loc
character used optionally when local is equal TRUE. Default: NULL

Value

list of list with names proper for DESCRIPTION file fields.

Examples

pacs_description(c("dplyr", "memoise"), version = c("0.8.1", "1.0.0"))
#> $dplyr
#> $dplyr$Type
#> [1] "Package"
#> 
#> $dplyr$Package
#> [1] "dplyr"
#> 
#> $dplyr$Title
#> [1] "A Grammar of Data Manipulation"
#> 
#> $dplyr$Version
#> [1] "0.8.1"
#> 
#> $dplyr$`Authors@R`
#> [1] "c(\nperson(\"Hadley\", \"Wickham\", , \"hadley@rstudio.com\", c(\"aut\", \"cre\"), comment = c(ORCID = \"0000-0003-4757-117X\")),\nperson(\"Romain\", \"Fran\\u00e7ois\", role = \"aut\", comment = c(ORCID = \"0000-0002-2444-4226\")),\nperson(\"Lionel\", \"Henry\", role = \"aut\"),\nperson(\"Kirill\", \"M\\u00fcller\", role = \"aut\", comment = c(ORCID = \"0000-0002-1416-3412\")),\nperson(\"RStudio\", role = c(\"cph\", \"fnd\"))\n)"
#> 
#> $dplyr$Description
#> [1] "A fast, consistent tool for working with data frame like objects,\nboth in memory and out of memory."
#> 
#> $dplyr$License
#> [1] "MIT + file LICENSE"
#> 
#> $dplyr$URL
#> [1] "http://dplyr.tidyverse.org, https://github.com/tidyverse/dplyr"
#> 
#> $dplyr$BugReports
#> [1] "https://github.com/tidyverse/dplyr/issues"
#> 
#> $dplyr$Depends
#> [1] "R (>= 3.2.0)"
#> 
#> $dplyr$Imports
#> [1] "assertthat (>= 0.2.1), glue (>= 1.3.1), magrittr (>= 1.5),\nmethods, pkgconfig (>= 2.0.2), R6 (>= 2.4.0), Rcpp (>= 1.0.1),\nrlang (>= 0.3.4), tibble (>= 2.1.1), tidyselect (>= 0.2.5),\nutils"
#> 
#> $dplyr$Suggests
#> [1] "bit64 (>= 0.9-7), callr (>= 3.2.0), covr (>= 3.2.1), DBI (>=\n1.0.0), dbplyr (>= 1.4.0), dtplyr (>= 0.0.3), ggplot2 (>=\n3.1.1), hms (>= 0.4.2), knitr (>= 1.22), Lahman (>= 6.0-0),\nlubridate (>= 1.7.4), MASS, mgcv (>= 1.8.23), microbenchmark\n(>= 1.4-6), nycflights13 (>= 1.0.0), rmarkdown (>= 1.12),\nRMySQL (>= 0.10.17), RPostgreSQL (>= 0.6-2), RSQLite (>=\n2.1.1), testthat (>= 2.1.1), withr (>= 2.1.2), broom (>=\n0.5.2), purrr (>= 0.3.2), readr (>= 1.3.1), crayon (>= 1.3.4)"
#> 
#> $dplyr$LinkingTo
#> [1] "BH (>= 1.69.0-1), plogr (>= 0.2.0), Rcpp (>= 1.0.1)"
#> 
#> $dplyr$VignetteBuilder
#> [1] "knitr"
#> 
#> $dplyr$Encoding
#> [1] "UTF-8"
#> 
#> $dplyr$LazyData
#> [1] "yes"
#> 
#> $dplyr$RoxygenNote
#> [1] "6.1.1"
#> 
#> $dplyr$NeedsCompilation
#> [1] "yes"
#> 
#> $dplyr$Packaged
#> [1] "2019-05-08 09:12:13 UTC; romainfrancois"
#> 
#> $dplyr$Author
#> [1] "Hadley Wickham [aut, cre] (<https://orcid.org/0000-0003-4757-117X>),\nRomain François [aut] (<https://orcid.org/0000-0002-2444-4226>),\nLionel Henry [aut],\nKirill Müller [aut] (<https://orcid.org/0000-0002-1416-3412>),\nRStudio [cph, fnd]"
#> 
#> $dplyr$Maintainer
#> [1] "Hadley Wickham <hadley@rstudio.com>"
#> 
#> $dplyr$Repository
#> [1] "CRAN"
#> 
#> $dplyr$`Date/Publication`
#> [1] "2019-05-14 12:20:03 UTC"
#> 
#> 
#> $memoise
#> $memoise$Encoding
#> [1] "UTF-8"
#> 
#> $memoise$Package
#> [1] "memoise"
#> 
#> $memoise$Title
#> [1] "Memoisation of Functions"
#> 
#> $memoise$Version
#> [1] "1.0.0"
#> 
#> $memoise$`Authors@R`
#> [1] "c(\nperson(\"Hadley\", \"Wickham\", , \"hadley@rstudio.com\", role = \"aut\"),\nperson(\"Jim\", \"Hester\", , \"jim.hester@rstudio.com\", role = c(\"aut\", \"cre\")),\nperson(\"Kirill\", \"Müller\", , \"krlmlr+r@mailbox.org\", role = \"aut\"))"
#> 
#> $memoise$Description
#> [1] "Cache the results of a function so that when you call it\nagain with the same arguments it returns the pre-computed value."
#> 
#> $memoise$URL
#> [1] "https://github.com/hadley/memoise"
#> 
#> $memoise$BugReports
#> [1] "https://github.com/hadley/memoise/issues"
#> 
#> $memoise$Imports
#> [1] "digest (>= 0.6.3)"
#> 
#> $memoise$Suggests
#> [1] "testthat"
#> 
#> $memoise$License
#> [1] "MIT + file LICENSE"
#> 
#> $memoise$RoxygenNote
#> [1] "5.0.1"
#> 
#> $memoise$NeedsCompilation
#> [1] "no"
#> 
#> $memoise$Packaged
#> [1] "2016-01-28 19:30:15 UTC; jhester"
#> 
#> $memoise$Author
#> [1] "Hadley Wickham [aut],\nJim Hester [aut, cre],\nKirill Müller [aut]"
#> 
#> $memoise$Maintainer
#> [1] "Jim Hester <jim.hester@rstudio.com>"
#> 
#> $memoise$Repository
#> [1] "CRAN"
#> 
#> $memoise$`Date/Publication`
#> [1] "2016-01-29 05:58:01"
#> 
#> 
pacs_description(c("dplyr", "memoise"), at = as.Date("2019-02-01"))
#> $dplyr
#> $dplyr$Type
#> [1] "Package"
#> 
#> $dplyr$Package
#> [1] "dplyr"
#> 
#> $dplyr$Title
#> [1] "A Grammar of Data Manipulation"
#> 
#> $dplyr$Version
#> [1] "0.7.8"
#> 
#> $dplyr$`Authors@R`
#> [1] "c(\nperson(\"Hadley\", \"Wickham\", , \"hadley@rstudio.com\", c(\"aut\", \"cre\"), comment = c(ORCID = \"0000-0003-4757-117X\")),\nperson(\"Romain\", \"Fran\\u00e7ois\", role = \"aut\", comment = c(ORCID = \"0000-0002-2444-4226\")),\nperson(\"Lionel\", \"Henry\", role = \"aut\"),\nperson(\"Kirill\", \"M\\u00fcller\", role = \"aut\", comment = c(ORCID = \"0000-0002-1416-3412\")),\nperson(\"RStudio\", role = c(\"cph\", \"fnd\"))\n)"
#> 
#> $dplyr$Description
#> [1] "A fast, consistent tool for working with data frame like objects,\nboth in memory and out of memory."
#> 
#> $dplyr$License
#> [1] "MIT + file LICENSE"
#> 
#> $dplyr$URL
#> [1] "http://dplyr.tidyverse.org, https://github.com/tidyverse/dplyr"
#> 
#> $dplyr$BugReports
#> [1] "https://github.com/tidyverse/dplyr/issues"
#> 
#> $dplyr$Depends
#> [1] "R (>= 3.1.2)"
#> 
#> $dplyr$Imports
#> [1] "assertthat (>= 0.2.0), bindrcpp (>= 0.2.0.9000), glue (>=\n1.1.1), magrittr (>= 1.5), methods, pkgconfig (>= 2.0.1), R6\n(>= 2.2.2), Rcpp (>= 0.12.19), rlang (>= 0.3.0), tibble (>=\n1.4.2), tidyselect (>= 0.2.3), utils"
#> 
#> $dplyr$Suggests
#> [1] "bit64 (>= 0.9.7), callr, covr (>= 3.0.1), DBI (>= 0.7.14),\ndbplyr (>= 1.2.0), dtplyr (>= 0.0.2), ggplot2 (>= 2.2.1), hms\n(>= 0.4.1), knitr (>= 1.19), Lahman (>= 3.0-1), lubridate,\nMASS, mgcv (>= 1.8.23), microbenchmark (>= 1.4.4), nycflights13\n(>= 0.2.2), rmarkdown (>= 1.8), RMySQL (>= 0.10.13),\nRPostgreSQL (>= 0.6.2), RSQLite (>= 2.0), testthat (>= 2.0.0),\nwithr (>= 2.1.1)"
#> 
#> $dplyr$LinkingTo
#> [1] "BH (>= 1.58.0-1), bindrcpp (>= 0.2.0.9000), plogr (>=\n0.1.10), Rcpp (>= 0.12.15)"
#> 
#> $dplyr$VignetteBuilder
#> [1] "knitr"
#> 
#> $dplyr$Encoding
#> [1] "UTF-8"
#> 
#> $dplyr$LazyData
#> [1] "yes"
#> 
#> $dplyr$RoxygenNote
#> [1] "6.1.0"
#> 
#> $dplyr$NeedsCompilation
#> [1] "yes"
#> 
#> $dplyr$Packaged
#> [1] "2018-11-09 21:23:40 UTC; kirill"
#> 
#> $dplyr$Author
#> [1] "Hadley Wickham [aut, cre] (<https://orcid.org/0000-0003-4757-117X>),\nRomain François [aut] (<https://orcid.org/0000-0002-2444-4226>),\nLionel Henry [aut],\nKirill Müller [aut] (<https://orcid.org/0000-0002-1416-3412>),\nRStudio [cph, fnd]"
#> 
#> $dplyr$Maintainer
#> [1] "Hadley Wickham <hadley@rstudio.com>"
#> 
#> $dplyr$Repository
#> [1] "CRAN"
#> 
#> $dplyr$`Date/Publication`
#> [1] "2018-11-10 07:30:02 UTC"
#> 
#> 
#> $memoise
#> $memoise$Encoding
#> [1] "UTF-8"
#> 
#> $memoise$Package
#> [1] "memoise"
#> 
#> $memoise$Title
#> [1] "Memoisation of Functions"
#> 
#> $memoise$Version
#> [1] "1.1.0"
#> 
#> $memoise$`Authors@R`
#> [1] "c(\nperson(\"Hadley\", \"Wickham\", , \"hadley@rstudio.com\", role = \"aut\"),\nperson(\"Jim\", \"Hester\", , \"jim.hester@rstudio.com\", role = c(\"aut\", \"cre\")),\nperson(\"Kirill\", \"Müller\", , \"krlmlr+r@mailbox.org\", role = \"aut\"),\nperson(\"Daniel\", \"Cook\", , \"danielecook@gmail.com\", role = \"aut\"))"
#> 
#> $memoise$Description
#> [1] "Cache the results of a function so that when you call it\nagain with the same arguments it returns the pre-computed value."
#> 
#> $memoise$URL
#> [1] "https://github.com/hadley/memoise"
#> 
#> $memoise$BugReports
#> [1] "https://github.com/hadley/memoise/issues"
#> 
#> $memoise$Imports
#> [1] "digest (>= 0.6.3)"
#> 
#> $memoise$Suggests
#> [1] "testthat, aws.s3, httr, covr"
#> 
#> $memoise$Additional_repositories
#> [1] "http://cloudyr.github.io/drat"
#> 
#> $memoise$License
#> [1] "MIT + file LICENSE"
#> 
#> $memoise$RoxygenNote
#> [1] "6.0.1"
#> 
#> $memoise$NeedsCompilation
#> [1] "no"
#> 
#> $memoise$Packaged
#> [1] "2017-04-20 17:35:47 UTC; jhester"
#> 
#> $memoise$Author
#> [1] "Hadley Wickham [aut],\nJim Hester [aut, cre],\nKirill Müller [aut],\nDaniel Cook [aut]"
#> 
#> $memoise$Maintainer
#> [1] "Jim Hester <jim.hester@rstudio.com>"
#> 
#> $memoise$Repository
#> [1] "CRAN"
#> 
#> $memoise$`Date/Publication`
#> [1] "2017-04-21 05:54:22 UTC"
#> 
#>