Packages health state at a specific Date or for a specific versions

using CRAN website to get a package version/versions used at a specific Date interval. A healthy package is a if it was published for more than 14 days. CRAN team gives around one week to resolved a package which gave errors under the check page.

pacs_health(pacs, versions = NULL, at = NULL)

Arguments

pacs
character vector packages names.

versions
character vector versions of packages. Default: NULL

at
Date old version of package. Default: NULL

Value

logical vector, TRUE if a package is healthy.

Note

Function will scrap two CRAN URLS. Works only with CRAN packages. Please as a courtesy to the R CRAN, don't overload their server by constantly using this function. Results are cached for 1 hour with memoise package, memory cache.

Examples

pacs_health(c("memoise"))
#> $memoise
#> [1] TRUE
#> 
pacs_health(c("memoise", "devtools"), versions = c("1.0.0", "2.4.0"))
#> $memoise
#> [1] TRUE
#> 
#> $devtools
#> [1] TRUE
#>