Packages life duration for a specific version

using CRAN website to get a package life duration for certain version or at a specific Date.

pacs_lifeduration(
  pacs,
  versions,
  source = c("crandb", "loop_crandb", "loop_cran"),
  lib.loc = .libPaths(),
  repos = biocran_repos()
)

Arguments

pacs

character vector packages names.

versions

character vector packages versions.

source

character one of c("crandb", "loop_crandb", "loop_cran"). The "crandb" works if less than getOption("pacs.crandb_limit") (currently 500) packages are looked for. Default: "crandb"

lib.loc

character vector. Is omitted for non NULL version. Default: .libPaths()

repos

character the base CRAN URL of the repository to use. Default "https://cran.rstudio.org"

Value

data.frame with two columns package name and life duration.

Note

Function will scrap two/tree CRAN URLS. Works only with CRAN packages. Results are cached for 30 minutes with memoise package, memory cache. The crandb R packages database is a part of METACRAN project, source: Csárdi G, Salmon M (2022). pkgsearch: Search and Query CRAN R Packages. https://github.com/r-hub/pkgsearch, https://r-hub.github.io/pkgsearch/.

Examples

if (FALSE) {
pacs_lifeduration(c("dplyr", "tidyr"), c("1.0.0", "1.2.0"))
pacs_lifeduration(c("dplyr", "tidyr"), c("1.0.0", "1.2.0"), source = "loop_cran")
}