Packages versions at a specific Date or a Date interval

using CRAN website to get packages version/versions used at a specific Date or a Date interval.

pacs_timemachine(pacs, at = NULL, from = NULL, to = NULL)

Arguments

pacs
character vector packages names.

at
Date old version of package. Default: NULL

from
Date new version of package. Default: NULL

to
Date CRAN URL. Default: NULL

Value

data.frame

Note

Function will scrap two CRAN URLS. Works only with CRAN packages. For bigger lists might need a few minutes. Please as a courtesy to the R CRAN, don't overload their server by constantly using this function. The base part of URL in the result is https://cran.r-project.org/src/contrib/.

Examples

pacs_timemachine(c("dplyr", "memoise"), from = as.Date("2018-06-30"), to = as.Date("2019-01-01"))
#> $dplyr
#>    Package Version   Released   Archived Life_Duration
#> 20   dplyr   0.7.6 2018-06-29 2018-10-16      109 days
#> 21   dplyr   0.7.7 2018-10-16 2018-11-10       25 days
#> 22   dplyr   0.7.8 2018-11-10 2019-02-14       96 days
#>                                 URL Size
#> 20 Archive/dplyr/dplyr_0.7.6.tar.gz 694K
#> 21 Archive/dplyr/dplyr_0.7.7.tar.gz 707K
#> 22 Archive/dplyr/dplyr_0.7.8.tar.gz 703K
#> 
#> $memoise
#>   Package Version   Released   Archived Life_Duration
#> 4 memoise   1.1.0 2017-04-21 2021-01-26     1376 days
#>                                    URL Size
#> 4 Archive/memoise/memoise_1.1.0.tar.gz  11K
#> 
pacs_timemachine(c("dplyr", "memoise"), at = Sys.Date())
#> $dplyr
#>     Package Version   Released Archived Life_Duration                URL Size
#> 110   dplyr   1.0.7 2021-06-18     <NA>       39 days dplyr_1.0.7.tar.gz <NA>
#> 
#> $memoise
#>   Package Version   Released Archived Life_Duration                  URL Size
#> 5 memoise   2.0.0 2021-01-26     <NA>      182 days memoise_2.0.0.tar.gz <NA>
#>