Validate specific packages

Checking the healthy of the specific packages.

pacs_validate(
  pacs,
  lib.loc = NULL,
  fields = c("Depends", "Imports", "LinkingTo"),
  lifeduration = FALSE,
  checkred = FALSE
)

Arguments

pacs
character vector packages names.

lib.loc
character. Default: NULL

fields
character vector with possible values c("Depends", "Imports", "LinkingTo", "Suggests"). Default: c("Depends", "Imports", "LinkingTo")

lifeduration
logical if to add life duration column, might take some time. Default: FALSE

checkred
logical if to add R CRAN check page status, any WARNING or ERROR will give TRUE. Default FALSE

Value

data.frame with 5/6/7 columns.

Package

character package names.

Version.expected.min

character expected by DESCRIPTION files minimal version. "" means not specified so the newest version.

Version.have

character installed package version.

version_status

numeric -1/0/1 which comes from utils::compareVersion function. 0 means that we have the same version as required by DESCRIPTION files. -1 means we have too low version installed, this is an error. 1 means we have higher version.

life_duration

(Optional) integer number of days package was released.

newest

logical if the installed version is the newest one.

checkred

(Optional) logical if the newest package contains any errors or warnings on CRAN check page.

Note

Version.expected.min column not count packages which are not a dependency for any package, so could not be find in DESCRIPTION files. When turn on the lifeduration and/or checkred options, calculations might be time consuming. Please as a courtesy to the R CRAN, don't overload their server by constantly using this function with lifeduration or checkred turned on. Results are cached for 1 hour with memoise package, memory cache.

Examples

pacs_validate(c("memoise", "rlang"))
#>   Package Version.expected.min Version.have version_status newest
#> 1  cachem                             1.0.5              1   TRUE
#> 2 fastmap                             1.1.0              1   TRUE
#> 3   rlang               0.4.10       0.4.11              1   TRUE