Validate the library.

Compare current and expected packages under .libPaths. Checking the healthy of the library, which packages are newest one. Optionally added life duration of each package.

lib_validate(
  lib.loc = NULL,
  fields = c("Depends", "Imports", "LinkingTo"),
  lifeduration = FALSE,
  is_red = FALSE
)

Arguments

lib.loc
character. Default: NULL

fields
character vector with possible values c("Depends", "Imports", "LinkingTo", "Suggests"). Default: c("Depends", "Imports", "LinkingTo")

lifeduration
logical if to add life duration column, might take some time. Default: FALSE

is_red
logical if to add R CRAN check page status, any WARNING or ERROR will give TRUE. Default FALSE

Value

data.frame with 5/7 columns Package Version.expected.min Version.have. "" means newest version.

Note

Version.expected.min column not count packages which are not a dependency for any package, so could not be find in DESCRIPTION files.

Examples

lib_validate()
#>         Package Version.expected.min Version.have version_status newest
#> 1    KernSmooth                 <NA>      2.23-20              1   TRUE
#> 2          MASS                            7.3-54              1   TRUE
#> 3        Matrix                             1.3-3              1  FALSE
#> 5             R                4.0.0        4.1.0              1     NA
#> 6            R6                2.2.0        2.5.0              1   TRUE
#> 7          Rcpp               0.11.0        1.0.6              1  FALSE
#> 8       askpass                               1.1              1   TRUE
#> 10    base64enc                             0.1-3              1   TRUE
#> 11         boot                 <NA>       1.3-28              1   TRUE
#> 12         brew                             1.0-6              1   TRUE
#> 13         brio                             1.1.2              1   TRUE
#> 14        bslib                           0.2.5.1              1   TRUE
#> 15       cachem                             1.0.5              1   TRUE
#> 16        callr                3.6.0        3.7.0              1   TRUE
#> 17        class                 <NA>       7.3-19              1   TRUE
#> 18          cli                2.4.0        3.0.1              1   TRUE
#> 19        clipr                0.3.0        0.7.1              1   TRUE
#> 20      cluster                 <NA>        2.1.2              1   TRUE
#> 21    codetools                 <NA>       0.2-18              1   TRUE
#> 22   commonmark                               1.7              1   TRUE
#> 24        cpp11                0.2.1        0.3.1              1   TRUE
#> 25       crayon                1.3.4        1.4.1              1   TRUE
#> 26  credentials                1.2.1        1.3.0              1  FALSE
#> 27         curl                3.0.0        4.3.2              1   TRUE
#> 29         desc                1.3.0        1.3.0              0   TRUE
#> 30     devtools                 <NA>        2.4.2              1   TRUE
#> 31      diffobj                             0.3.4              1   TRUE
#> 32       digest                            0.6.27              1   TRUE
#> 33      downlit                        0.2.1.9000              1  FALSE
#> 34     ellipsis                0.3.2        0.3.2              0   TRUE
#> 35     evaluate                 0.13         0.14              1   TRUE
#> 36        fansi                0.4.0        0.5.0              1   TRUE
#> 37      fastmap                             1.1.0              1   TRUE
#> 38      foreign                 <NA>       0.8-81              1   TRUE
#> 39           fs                1.5.0        1.5.0              0   TRUE
#> 40         gert                1.0.2        1.3.1              1   TRUE
#> 41           gh                1.2.0        1.3.0              1   TRUE
#> 42     gitcreds                             0.1.1              1   TRUE
#> 43         glue                1.3.0        1.4.2              1   TRUE
#> 47        highr                               0.9              1   TRUE
#> 48    htmltools                0.5.1      0.5.1.1              1   TRUE
#> 49         httr                1.4.2        1.4.2              0   TRUE
#> 50          ini                             0.3.1              1   TRUE
#> 51    jquerylib                0.1.3        0.1.4              1   TRUE
#> 52     jsonlite                             1.7.2              1   TRUE
#> 53        knitr                 1.22         1.33              1   TRUE
#> 54      lattice                           0.20-44              1   TRUE
#> 55    lifecycle                1.0.0        1.0.0              0   TRUE
#> 56     magrittr                  1.5        2.0.1              1   TRUE
#> 57     markdown                               1.1              1   TRUE
#> 58      memoise                2.0.0        2.0.0              0   TRUE
#> 60         mgcv                 <NA>       1.8-35              1  FALSE
#> 61         mime                  0.3         0.11              1   TRUE
#> 62         nlme               3.1-64      3.1-152              1   TRUE
#> 63         nnet                 <NA>       7.3-16              1   TRUE
#> 64      openssl                1.4.1        1.4.4              1   TRUE
#> 65      packrat                 <NA>        0.6.0              1   TRUE
#> 66         pacs                 <NA>        0.2.3              1     NA
#> 68       pillar                1.6.0        1.6.1              1   TRUE
#> 69     pkgbuild                1.2.0        1.2.0              0   TRUE
#> 70    pkgconfig                             2.0.3              1   TRUE
#> 71      pkgdown                 <NA>   1.6.1.9001              1  FALSE
#> 72      pkgload                1.2.1        1.2.1              0   TRUE
#> 73       praise                             1.0.0              1   TRUE
#> 74  prettyunits                             1.1.1              1   TRUE
#> 75     processx                3.5.0        3.5.2              1   TRUE
#> 76           ps                1.3.4        1.6.0              1   TRUE
#> 77        purrr                0.3.3        0.3.4              1   TRUE
#> 78         ragg                             1.1.3              1   TRUE
#> 79     rappdirs                             0.3.3              1   TRUE
#> 80    rcmdcheck                1.3.3        1.3.3              0   TRUE
#> 81     rematch2                             2.1.2              1   TRUE
#> 82      remotes                2.3.0        2.4.0              1   TRUE
#> 83        rlang               0.4.10       0.4.11              1   TRUE
#> 84    rmarkdown             1.1.9007          2.9              1   TRUE
#> 85     roxygen2                7.1.1        7.1.1              0   TRUE
#> 86        rpart                 <NA>       4.1-15              1   TRUE
#> 87    rprojroot                  1.2        2.0.2              1   TRUE
#> 88   rstudioapi                 0.13         0.13              0   TRUE
#> 89    rversions                2.0.2        2.1.1              1   TRUE
#> 90         sass                0.4.0        0.4.0              0   TRUE
#> 91  sessioninfo                1.1.1        1.1.1              0   TRUE
#> 92      spatial                 <NA>       7.3-14              1   TRUE
#> 96      stringi                1.1.7        1.6.2              1  FALSE
#> 97      stringr                1.2.0        1.4.0              1   TRUE
#> 98     survival                 <NA>       3.2-11              1   TRUE
#> 99          sys                  2.1          3.4              1   TRUE
#> 100 systemfonts                1.0.0        1.0.2              1   TRUE
#> 102    testthat                3.0.2        3.0.4              1   TRUE
#> 103 textshaping                0.3.0        0.3.5              1   TRUE
#> 104      tibble                             3.1.3              1   TRUE
#> 105     tinytex                 0.31         0.32              1   TRUE
#> 107     usethis                2.0.1        2.0.1              0   TRUE
#> 108        utf8                1.1.0        1.2.2              1   TRUE
#> 110       vctrs                0.3.8        0.3.8              0   TRUE
#> 111       waldo                0.2.4        0.2.5              1   TRUE
#> 112     whisker                               0.4              1   TRUE
#> 113       withr                2.4.1        2.4.2              1   TRUE
#> 114        xfun                 0.23         0.24              1   TRUE
#> 115        xml2                1.3.1        1.3.2              1   TRUE
#> 116       xopen                             1.0.0              1   TRUE
#> 117        yaml               2.1.19        2.2.1              1   TRUE
#> 118         zip                2.1.0        2.2.0              1   TRUE