Compare current and expected packages under .libPaths.

Checking the healthy of the libarary.

lib_validate(lib.loc = NULL, fields = c("Depends", "Imports", "LinkingTo"))

Arguments

lib.loc
character. Default: NULL

fields
character vector with possible values c("Depends", "Imports", "LinkingTo", "Suggests"). Default: c("Depends", "Imports", "LinkingTo")

Value

data.frame with 3 columns Package Version.expected.min Version.have. "" means newest version.

Note

Version.expected.min column not count packages which are not a dependency for any package, so could not be find in DESCRIPTION files.

Examples

lib_validate()
#>         Package Version.expected.min Version.have
#> 1    KernSmooth                 <NA>      2.23-20
#> 2          MASS                            7.3-54
#> 3        Matrix                             1.3-3
#> 4            NA                              <NA>
#> 5             R                4.0.0         <NA>
#> 6            R6                2.2.0        2.5.0
#> 7          Rcpp               0.11.0        1.0.6
#> 8       askpass                               1.1
#> 9          base                 <NA>        4.1.0
#> 10    base64enc                             0.1-3
#> 11         boot                 <NA>       1.3-28
#> 12         brew                             1.0-6
#> 13         brio                             1.1.2
#> 14        bslib                           0.2.5.1
#> 15       cachem                             1.0.5
#> 16        callr                3.6.0        3.7.0
#> 17        class                 <NA>       7.3-19
#> 18          cli                2.4.0        3.0.0
#> 19        clipr                0.3.0        0.7.1
#> 20      cluster                 <NA>        2.1.2
#> 21    codetools                 <NA>       0.2-18
#> 22   commonmark                               1.7
#> 23     compiler                             4.1.0
#> 24        cpp11                0.2.1        0.3.1
#> 25       crayon                1.3.4        1.4.1
#> 26  credentials                1.2.1        1.3.0
#> 27         curl                3.0.0        4.3.2
#> 28     datasets                 <NA>        4.1.0
#> 29         desc                1.3.0        1.3.0
#> 30     devtools                 <NA>        2.4.2
#> 31      diffobj                             0.3.4
#> 32       digest                            0.6.27
#> 33      downlit                        0.2.1.9000
#> 34     ellipsis                0.3.2        0.3.2
#> 35     evaluate                 0.13         0.14
#> 36        fansi                0.4.0        0.5.0
#> 37      fastmap                             1.1.0
#> 38      foreign                 <NA>       0.8-81
#> 39           fs                1.5.0        1.5.0
#> 40         gert                1.0.2        1.3.1
#> 41           gh                1.2.0        1.3.0
#> 42     gitcreds                             0.1.1
#> 43         glue                1.3.0        1.4.2
#> 44    grDevices                             4.1.0
#> 45     graphics                             4.1.0
#> 46         grid                             4.1.0
#> 47        highr                               0.9
#> 48    htmltools                0.5.1      0.5.1.1
#> 49         httr                1.4.2        1.4.2
#> 50          ini                             0.3.1
#> 51    jquerylib                0.1.3        0.1.4
#> 52     jsonlite                             1.7.2
#> 53        knitr                 1.22         1.33
#> 54      lattice                           0.20-44
#> 55    lifecycle                1.0.0        1.0.0
#> 56     magrittr                  1.5        2.0.1
#> 57     markdown                               1.1
#> 58      memoise                2.0.0        2.0.0
#> 59      methods                             4.1.0
#> 60         mgcv                 <NA>       1.8-35
#> 61         mime                  0.3         0.11
#> 62         nlme               3.1-64      3.1-152
#> 63         nnet                 <NA>       7.3-16
#> 64      openssl                1.4.1        1.4.4
#> 65         pacs                 <NA>        0.1.2
#> 66     parallel                 <NA>        4.1.0
#> 67       pillar                1.6.0        1.6.1
#> 68     pkgbuild                1.2.0        1.2.0
#> 69    pkgconfig                             2.0.3
#> 70      pkgdown                 <NA>   1.6.1.9001
#> 71      pkgload                1.2.1        1.2.1
#> 72       praise                             1.0.0
#> 73  prettyunits                             1.1.1
#> 74     processx                3.5.0        3.5.2
#> 75           ps                1.3.4        1.6.0
#> 76        purrr                0.3.3        0.3.4
#> 77         ragg                             1.1.3
#> 78     rappdirs                             0.3.3
#> 79    rcmdcheck                1.3.3        1.3.3
#> 80     rematch2                             2.1.2
#> 81      remotes                2.3.0        2.4.0
#> 82        rlang               0.4.10       0.4.11
#> 83    rmarkdown             1.1.9007          2.9
#> 84     roxygen2                7.1.1        7.1.1
#> 85        rpart                 <NA>       4.1-15
#> 86    rprojroot                  1.2        2.0.2
#> 87   rstudioapi                 0.13         0.13
#> 88    rversions                2.0.2        2.1.1
#> 89         sass                0.4.0        0.4.0
#> 90  sessioninfo                1.1.1        1.1.1
#> 91      spatial                 <NA>       7.3-14
#> 92      splines                             4.1.0
#> 93        stats                             4.1.0
#> 94       stats4                 <NA>        4.1.0
#> 95      stringi                1.1.7        1.6.2
#> 96      stringr                1.2.0        1.4.0
#> 97     survival                 <NA>       3.2-11
#> 98          sys                  2.1          3.4
#> 99  systemfonts                1.0.0        1.0.2
#> 100       tcltk                 <NA>        4.1.0
#> 101    testthat                3.0.2        3.0.4
#> 102 textshaping                0.3.0        0.3.5
#> 103      tibble                             3.1.2
#> 104     tinytex                 0.31         0.32
#> 105       tools                             4.1.0
#> 106     usethis                2.0.1        2.0.1
#> 107        utf8                1.1.0        1.2.1
#> 108       utils                             4.1.0
#> 109       vctrs                0.3.8        0.3.8
#> 110       waldo                0.2.4        0.2.5
#> 111     whisker                               0.4
#> 112       withr                2.4.1        2.4.2
#> 113        xfun                 0.23         0.24
#> 114        xml2                1.3.1        1.3.2
#> 115       xopen                             1.0.0
#> 116        yaml               2.1.19        2.2.1
#> 117         zip                2.1.0        2.2.0