package NAMESPACE file

CRAN package NAMESPACE file taken locally or remotely from GITHUB CRAN mirror or CRAN website.

pac_namespace(
  pac,
  version = NULL,
  at = NULL,
  local = FALSE,
  lib.loc = .libPaths(),
  repos = "https://cran.rstudio.com/"
)

Arguments

pac

character a package name.

version

character package version, By default the newest version in taken if failed tried to give local one if installed. Default: NULL

at

Date. Default: NULL

local

logical if to use local library. Default: FALSE

lib.loc

character vector, used optionally when local is equal TRUE. Default: .libPaths()

repos

character the base URL of the CRAN repository to use. Used only for the validation. Default https://cran.rstudio.com/

Value

list with names proper for NAMESPACE file, the same as format as returned by base::parseNamespaceFile.

Note

Results are cached for 1 hour with memoise package. This function is mainly built under source code from base::parseNamespaceFile.

Examples

if (FALSE) {
pac_namespace("dplyr", version = "0.8.0")
pac_namespace("dplyr", at = as.Date("2019-02-01"))
pac_namespace("memoise", local = TRUE)
}